+++ .gitignore
@@ -8,6 +8,11 @@ node_modules/
 # Ignore logs
 logs/
 *.log
+*.log.global
+
+# Ignore video/image exports
+*.png
+*.mp4
 
 # Ignore OS generated files
 .DS_Store
