+++ src/multi_robot_multi_goal_planning/planners/prioritized_planner.py
@@ -49,13 +49,44 @@ from multi_robot_multi_goal_planning.problems.util import path_cost
 
 import logging
 
+"""
+NOTES (Liam):
+=============================================================================================
+SUMMARY
+
+Core idea: 
+- Plan one task at a time instead of all simultaneously (computationally expensive)
+- Robot involved in the current task are planned jointly 
+- All other robots (with already planned paths) are treated as moving obstacles
+
+Key steps:
+1. Get task sequence: Determine which task to do in which order (from oracle/task planner)
+2. For each task in sequence:
+    - Identify which robots are involved
+    - Plan their paths while avoiding robots that already have plans
+    - Add an "escape path" to move robots to safe poses afterward
+3. Iterate: Try different task sequences to find better solutions
+=============================================================================================
+"""
 logger = logging.getLogger(__name__)
 
 TimedPath = namedtuple("TimedPath", ["time", "path"])
 Path = namedtuple("Path", ["path", "task_index"])
 
+"""
+NOTES (Liam):
+Remove later also comments within code, keep for now for better understanding
+=============================================================================================
+MULTIROBOTPATH
 
+- Stores the complete planned paths for all robots over time, built incrementally as robots are
+  planned one by one in prioritized planning
+- Answers questions: where robot X at time t? what mode is active at time t?
+=============================================================================================
+"""
 class MultiRobotPath:
+    # Allocate space for fixed set of attributes (unlike __dict__)
+    # Memory saving & faster attribute access 
     __slots__ = [
         "robots",
         "paths",
@@ -81,55 +112,64 @@ class MultiRobotPath:
         for i, r in enumerate(robots):
             self.robot_ids[r] = i
 
-        # indicates which mode is active after a certain time (until the next one)
+        # Indicates which mode is active after a certain time (until the next one)
         self.timed_mode_sequence = [(0, self.m0)]
         self.times = [0]
 
         for r in robots:
             self.paths[r] = []
-
+    
     def get_mode_at_time(self, t: float) -> Mode:
-        """
-        Iterate over all the stored modes and give the one back that we are in at a certain time.
-        """
-
+        """Return active mode at time t"""
         if t >= self.times[-1]:
             return self.timed_mode_sequence[-1][1]
 
-        idx = bisect.bisect_right(self.times, t)
+        idx = bisect.bisect_right(self.times, t) # ([0,10,20,30],15) -> 2
 
-        return self.timed_mode_sequence[idx - 1][1]
+        return self.timed_mode_sequence[idx - 1][1] # 2 = mode_B
 
     # @profile # run with kernprof -l examples/run_planner.py [your environment] [your flags]
     def get_robot_poses_at_time(self, robots: List[str], t: float):
+        """
+        Calculates the exact configuration of a given list of robots at a specific time
+        """
         poses = []
 
         for r in robots:
             i = self.robot_ids[r]
 
+            # CASE 1: robot has NO path (not planned yet) -> stay at initial position
             if not self.paths[r]:
                 poses.append(self.q0_split[i])
             else:
+                # CASE 2: query time AFTER all paths END -> return last pose of last path segment
+                # Why? Robot stays stationary after completing all tasks
                 if t >= self.paths[r][-1].path.time[-1]:
                     pose = self.paths[r][-1].path.path[-1]
+                
+                # CASE 3: query time BEFORE all paths START -> return initial pose
                 elif t <= self.paths[r][0].path.time[0]:
                     pose = self.q0_split[i]
+                
+                # CASE 4: query time WITHIN the path -> find which segment + interpolate within
                 else:
                     path_end_times = [rp.path.time[-1] for rp in self.paths[r]]
-                    segment_idx = bisect.bisect_left(path_end_times, t)
-
+                    segment_idx = bisect.bisect_left(path_end_times, t) # Path segment time falls into
                     rp = self.paths[r][segment_idx]
-
+                    
+                    # 4A. query time == segment end time -> return last pose of that segment
                     if t == rp.path.time[-1] and segment_idx < len(self.paths[r]) - 1:
                         # If 't' is exactly the end time of a segment that is not the very last segment,
                         # we take the last pose of *that* segment.
                         pose = rp.path.path[-1]
+                    
+                    # 4B. query time IN segment -> linear interpolation within segment
                     elif rp.path.time[0] <= t < rp.path.time[-1]:
-                        # 't' is within the current segment (exclusive of the end time).
+                        # 't' is within the current segment (exclusive of the end time)
                         p = rp.path.path
                         time_points = rp.path.time
 
-                        # Find the specific time interval within the segment
+                        # Find the SPECIFIC time interval within the segment
                         k = bisect.bisect_right(time_points, t) - 1
 
                         time_k = time_points[k]
@@ -139,6 +179,8 @@ class MultiRobotPath:
                         td = time_kn - time_k
                         q0 = p[k]
                         pose = q0 + (t - time_k) / td * (p[k + 1] - q0)
+                    
+                    # 4C. edge case handling
                     else:
                         if (
                             segment_idx > 0
@@ -156,10 +198,12 @@ class MultiRobotPath:
         return poses
 
     def get_end_times(self, robots: List[str]):
+        """Returns dict of when each robot's last path ends"""
         end_times = {}
 
         for r in robots:
             if len(self.paths[r]) > 0:
+                # Get the last time of the last path segment of robot r
                 end_times[r] = self.paths[r][-1].path.time[-1]
             else:
                 end_times[r] = 0
@@ -167,14 +211,15 @@ class MultiRobotPath:
         return end_times
 
     def get_non_escape_end_times(self, robots: List[str]):
+        """Get end times excluding escape plans (task_index == -1)"""
         end_times = {}
 
         for r in robots:
             if len(self.paths[r]) > 0:
-                if self.paths[r][-1].task_index == -1:
-                    end_times[r] = self.paths[r][-2].path.time[-1]
+                if self.paths[r][-1].task_index == -1: # Escape path
+                    end_times[r] = self.paths[r][-2].path.time[-1] # [-2] is last non-escape path segment
                 else:
-                    end_times[r] = self.paths[r][-1].path.time[-1]
+                    end_times[r] = self.paths[r][-1].path.time[-1] # [-1] is last when there's no escape path segment
             else:
                 end_times[r] = 0
 
@@ -187,9 +232,25 @@ class MultiRobotPath:
         next_mode: Optional[Mode],
         is_escape_path: bool = False,
     ):
+        """
+        Add a new path segment for one or more robots
+        
+        Example: 
+        - Robot 1 finishes task, picks up object -> mode changes
+          path = plan_to_goal...
+          multi_robot_path.add_path(robot1, path, next_mode=mode_with_obj, is_escape=False)
+        - Robot 2 does temporary escape motion -> mode unchanged
+          escape_path = plan_escape...
+          multi_robot_path.add_path(robot2, escape_path, next_mode=None, is_escape=True)
+        """
         logger.debug("adding path to multi-robot-path")
+
+        # TODO (Liam)
+        # Add final_time=0.0 here and in the loop track latest finish time among involved robtos
+        # and not just the final_time of the LAST robot (like it is implemented now..) 
+
         for r in robots:
-            # get robot-path from the original path
+            # Get robot-path from the original path
             subpath = Path(
                 task_index=path.task_index,
                 path=path.path[r],
@@ -197,22 +258,26 @@ class MultiRobotPath:
             self.paths[r].append(subpath)
 
             final_time = path.path[r].time[-1]
+            # final_time = max(final_time, path.path[r].time[-1]) # TODO (Liam)
             logger.debug("max_time of path:", final_time)
+            # logger.debug("max_time of path for %s: %s", r, final_time) # TODO (Liam)
 
-        # constructing the mode-sequence:
-        # this is done simply by adding the next mode to the sequence
+        # Constructing the mode-sequence:
+        # This is done simply by adding the next mode to the sequence
         if not is_escape_path:
             self.timed_mode_sequence.append((final_time, next_mode))
             self.times.append(final_time)
 
     def remove_final_escape_path(self, robots: List[str]):
+        """Remove escape path that are no longer needed (if better plan is found)"""
         for r in robots:
             if not self.paths[r]:
                 continue
             if self.paths[r][-1].task_index == -1:
-                self.paths[r] = self.paths[r][:-1]
+                self.paths[r] = self.paths[r][:-1] # Keep all but last (remove escape)
 
     def get_final_time(self):
+        """Get the makespan (when the last robot finishes)"""
         T = 0
         for k, v in self.paths.items():
             if len(v) > 0:
@@ -221,16 +286,28 @@ class MultiRobotPath:
         return T
 
     def get_final_non_escape_time(self):
+        """Get the makespan excluding escape paths"""
         T = 0
         for k, v in self.paths.items():
             if len(v) > 0:
-                if v[-1].task_index != -1:
+                if v[-1].task_index != -1: # No escape paths
                     T = max(T, v[-1].path.time[-1])
-                else:
+                else: # Don't consider escape paths
                     T = max(T, v[-2].path.time[-1])
-
         return T
 
+    # TODO (Liam) Check for active skill at time t (easier inspecting mode than tracking time ranges..)    
+    def mode_has_skill_for_robot(self, env: BaseProblem, t: float, robot: str) -> bool:
+        """
+        Check if robot has an active skill at time t by inspecting the mode
+        """
+        mode = self.get_mode_at_time(t)
+        robot_idx = env.robot_ids[robot]
+        task_id = mode.task_ids[robot_idx]
+        if task_id is None:
+            return False
+        task = env.tasks[task_id]
+        return getattr(task, "skill", None) is not None
 
 def display_multi_robot_path(env: rai_env, path: MultiRobotPath, blocking=True):
     T = path.get_final_time()
@@ -249,25 +326,35 @@ def display_multi_robot_path(env: rai_env, path: MultiRobotPath, blocking=True):
             logger.info(f"Collision at time {t}")
             env.show(blocking=True)
 
-        # env.show(blocking=False)
-
         time.sleep(0.01)
 
+"""
+NOTES (Liam):
+Remove later also comments within code, keep for now for better understanding
+=============================================================================================
+NODE and TREE
 
+Node:
+- Single space-time (q,t) point in search tree
+- Tree links: parent node + list of children nodes
+
+Tree:
+- Manages nodes with space-time metric that, 1) respects temporal causality, 2) enforces velocity
+  limits, and 3) accounts for multi-robot coordination timing
+=============================================================================================
+"""
 class Node:
     def __init__(self, t: float, q: Configuration):
         self.t = t
         self.q = q
-
         self.parent = None
         self.children = []
 
-
 class Tree:
     def __init__(self, start: Optional[Node], reverse=False):
-        self.reverse = reverse
+        self.reverse = reverse # For bidirectional search
 
-        if start is not None:
+        if start is not None: # Checks if tree initialized with root
             self.nodes = [start]
             self.configs = [[start.q[i] for i in range(start.q.num_agents())]]
         else:
@@ -277,45 +364,70 @@ class Tree:
         self.batch_config_dist_fun = batch_config_dist
 
         self.prev_plans = None
-        self.robots = None
+        self.robots = None # Depending on task, contains 1 or more robots ()
 
         self.gamma = 0.7
 
     def batch_dist_fun(self, n1: Node, n2: List[Node], v_max):
+        """
+        Space time distance metric
+        Calculates the "distance" from a query node n1 (random sample) to a list of existing
+        tree nodes n2 (n). Filters out physically impossible connections
+        """
+        # Step 1: Time direction check
         ts = np.array([1.0 * n.t for n in n2])
-
         if self.reverse:
-            t_diff = ts - n1.t
-
+            t_diff = ts - n1.t # Reverse (backward) tree (parent.t > child.t) -> t_diff < 0
         else:
-            t_diff = n1.t - ts
+            t_diff = n1.t - ts # Forward tree (parent.t < child.t) -> t_diff > 0
 
+        # Step 2: Velocity feasibility check
         q_dist_for_vel = self.batch_config_dist_fun(n1.q, [n.q for n in n2])
-        v = q_dist_for_vel / t_diff
+        v = q_dist_for_vel / t_diff # Required velocities to connect n1 with n2's
 
+        # Mask for infeasible connections (can't violate physics and causality)
         mask = (t_diff < 0) | (abs(v) > v_max)
 
+        """
+        1. Get release times for each robot being planned (time when it finishes previously assigned path)
+        2. Prepare intermediate waypoints for each (n1,n2) pair 
+        3. Iterate through nearest nodes and robots and compute distance between nodes
+        3A. If segment crosses moment where robot ends prior task -> p = "release point"
+        3B. Else -> p = new proposed point
+        4. Compute distances
+        """
+        # Step 3A: Planning for multiple robots with different end times
         if len(self.robots) > 1:
             end_times = self.prev_plans.get_end_times(self.robots)
-
             intermediate_poses_arr = np.zeros((len(n2), len(self.nodes[0].q.state())))
+            
+            # Iterate through every node of tree n2
+            # Check if the connection n to n1 crosses the robot's end_time
             for i in range(len(n2)):
                 n = n2[i]
                 offset = 0
                 for j, r in enumerate(self.robots):
                     if n1.t > end_times[r] and n.t < end_times[r]:
+                        # Robot locked to previous path (segment n-n1 crosses robot's release time)
+                        # Timeline: n --(old path)-- end_times[r] --(new path)-- n1 
+                        # Set intermediate point p to release pose (very last pose from robot's old path)
                         p = self.prev_plans.get_robot_poses_at_time([r], n1.t)[0]
                     else:
-                        # p = n.q[j]
+                        # Robot is already free (finished previous task)
+                        # OR robot still busy (for entire duration) 
+                        # Set intermediate point p to start point n.q 
                         p = self.configs[i][j]
 
                     dim = len(p)
                     intermediate_poses_arr[i, offset : offset + dim] = 1.0 * p
                     offset += dim
-
+            
+            # Distance from sample to intermediate points 
             q_dist_to_inter = self.batch_config_dist_fun(
                 n1.q, intermediate_poses_arr, "max"
             )
+
+            # Distance from intermediate points back to actual tree nodes n2
             q_dist_from_inter = batch_config_cost(
                 intermediate_poses_arr,
                 np.array([n.q.state() for n in n2]),
@@ -323,105 +435,45 @@ class Tree:
                 tmp_agent_slice=n.q._array_slice,
             )
 
+            # Final weighted space-time distance 
             dist = (q_dist_from_inter + q_dist_to_inter) * self.gamma + (
-                1 - self.gamma
-            ) * t_diff
+                1 - self.gamma) * t_diff
 
+        # Step 3B: Planning for single robot
         else:
+            # Weighted straight space-time metric (distance) from n1 to each node in n2
             q_dist = self.batch_config_dist_fun(n1.q, [n.q for n in n2], "max")
             dist = q_dist * self.gamma + (1 - self.gamma) * t_diff
 
+        # Infeasible connections -> mark with infinite distance
         dist[mask] = np.inf
 
         return dist
 
-        # print(q_diff)
-
     def get_nearest_neighbor(self, node: Node, v_max) -> Optional[Node]:
+        """Find closest node in the tree to connect a new sample to"""
         batch_dists = self.batch_dist_fun(node, self.nodes, v_max)
         batch_idx = np.argmin(batch_dists)
 
         if np.isinf(batch_dists[batch_idx]):
             return None
-
-        # print(batch_dists[batch_idx], config_dist(node.q, self.nodes[batch_idx].q), node.t - self.nodes[batch_idx].t)
-
-        # print('time', node.t)
-
-        # if len(self.nodes) % 50 == 0 and len(self.nodes) > 0:
-        # if len(self.nodes[0].q.state()) == 6:
-        #     to_plt = [n.q.state()[3:5] for n in self.nodes]
-        # else:
-        #     to_plt = [n.q.state()[:2] for n in self.nodes]
-
-        # fig = plt.figure(figsize=(12, 12))
-        # ax = fig.add_subplot(projection='3d')
-
-        # ax.scatter([a[0] for a in to_plt], [a[1] for a in to_plt], [a.t for a in self.nodes], c=batch_dists, cmap=mpl.colormaps["binary"])
-        # ax.scatter(node.q.state()[3], node.q.state()[4], node.t, marker='x')
-        # ax.scatter(self.nodes[batch_idx].q.state()[3], self.nodes[batch_idx].q.state()[4], self.nodes[batch_idx].t, color='red')
-
-        # for i, n in enumerate(self.nodes):
-        #     k = 2
-        #     node_dists = self.batch_dist_fun(n, self.nodes, v_max)
-        #     ind = np.argpartition(node_dists, k)[:k]
-
-        #     for j in ind:
-        #         if not np.isinf(node_dists[j]):
-        #             ax.plot([n.q.state()[3], self.nodes[j].q.state()[3]], [n.q.state()[4], self.nodes[j].q.state()[4]], [n.t, self.nodes[j].t], color='black')
-
-        # ax.set_xlim([-2, 2])
-        # ax.set_ylim([-2, 2])
-
-        # plt.show()
-
-        # fig = plt.figure(figsize=(12, 12))
-        # ax = fig.add_subplot(projection='3d')
-        # for i, n in enumerate(self.nodes):
-        #     if n.parent is not None:
-        #         ax.plot([n.q.state()[0], n.parent.q.state()[0]], [n.q.state()[1], n.parent.q.state()[1]], [n.t, n.parent.t], color='black')
-
-        # ax.set_xlim([-2, 2])
-        # ax.set_ylim([-2, 2])
-
-        # ax.set_xlabel('x')
-
-        # ax.scatter(self.nodes[0].q.state()[0], self.nodes[0].q.state()[1], self.nodes[0].t, color='red')
-
-        # plt.show()
-
+        
         return self.nodes[batch_idx]
 
-        # best_node = None
-        # best_dist = None
-        # for n in self.nodes:
-        #     qd = config_dist(node.q, n.q)
-        #     td = node.t - n.t
-
-        #     if td < 0:
-        #         continue
-
-        #     if qd / td > v_max:
-        #         continue
-
-        #     if best_dist is None or qd < best_dist:
-        #         best_node = n
-        #         best_dist = qd
-
-        # return best_node
-
     def get_near_neighbors(self, node: Node, k: int, v_max) -> List[Node]:
+        """Find closest k-nodes in the tree to connect new sample to"""
         node_list = self.nodes
         dists = self.batch_dist_fun(node, self.nodes, v_max)
 
-        k_clip = min(k, len(node_list) - 1)
-        topk = np.argpartition(dists, k_clip)[:k_clip]
-        topk[np.argsort(dists[topk])]
+        k_clip = min(k, len(node_list) - 1) # Conservative "-1" to avoid returning the node itself if query node already in dataset
+        topk = np.argpartition(dists, k_clip)[:k_clip] # Sorting in O(N) with k_clip smallest moved to front
+        topk[np.argsort(dists[topk])] # TODO dead code -> creates new sorted array, needs assignment: topk = ...
 
-        best_nodes = [node_list[i] for i in topk]
+        best_nodes = [node_list[i] for i in topk] # TODO here topk is correct but UNORDERED set of k neighbours
         return best_nodes
 
     def add_node(self, new_node: Node, parent: Optional[Node]) -> None:
+        """Add a node to the tree and establish parent-child relationship"""
         node_list = self.nodes
         node_list.append(new_node)
 
@@ -431,7 +483,19 @@ class Tree:
             new_node.parent = parent
             parent.children.append(new_node)
 
+"""
+NOTES (Liam):
+Remove later also comments within code, keep for now for better understanding
+=============================================================================================
+COLLISION CHECK
 
+Collision free with moving obstacles:
+- Check if the configuration is collision-free at time t, consider other robots as moving obstacles
+
+Edge collision free with moving obstacles:
+- Check if an entire edge (path segment) is collision free over time 
+=============================================================================================
+"""
 # @profile # run with kernprof -l examples/run_planner.py [your environment] [your flags]
 def collision_free_with_moving_obs(
     env: BaseProblem,
@@ -443,70 +507,32 @@ def collision_free_with_moving_obs(
     robots,
     other_robots,
 ):
+    """
+    Collision check with robots (active and others)
+    """
     logger.debug(f"coll check for {robots} at time {t}")
     logger.debug("Config %s", q)
 
+    # Step 1: Get mode (what objects are attached, etc.) at time t
     mode = prev_plans.get_mode_at_time(t)
-    # env.set_to_mode(mode)
 
+    # Step 2: Get positions of other robots at time t from their existing plans 
     robot_poses = prev_plans.get_robot_poses_at_time(other_robots, t)
+    
+    # Step 3: Build complete configuration: other robots + robots being planned
     for i, r in enumerate(other_robots):
-        q_buffer[env.robot_idx[r]] = robot_poses[i]
-    # if other_robots:
-    # q_buffer[np.array([env.robot_idx[r] for r in other_robots])] = np.array(robot_poses).flatten()
+        q_buffer[env.robot_idx[r]] = robot_poses[i] # Other robots
 
     offset = 0
     for r in robots:
         dim = env.robot_dims[r]
-        q_buffer[env.robot_idx[r]] = q[offset : offset + dim]
+        q_buffer[env.robot_idx[r]] = q[offset : offset + dim] # Robots being planned
         offset += dim
 
-    # env.C.setJointState(q_buffer)
-
-    # if env.is_collision_free_np(q_buffer, mode):
+    # Step 4: Check collision
     if env.is_collision_free(env.start_pos.from_flat(q_buffer), mode):
-        # if t > 85:
-        #     env.show(blocking=False)
         return True
-    
-    # if t > 85:
-    #     env.show(blocking=False)
-        
-    # # involves_robot_we_plan_for = False
-    # colls = env.C.getCollisions()
-    # for c in colls:
-    #     for r in robots:
-    #         if (r in c[1] or r in c[0]) or ("obj" in c[0] or "obj" in c[1]):
-    #             # involves_robot_we_plan_for = True
-    #             return False
-    #         # else:
-    #         #     env.C.view(True)
-    #         #     print(c)
-    # env.C.view(True)
-    # # print(c)
-
-    # return True
-
-    # if c[2] < 0:
-    #     print(c)
-
-    # colls = env.C.getCollisions()
-    # for c in colls:
-    #     for r in robots:
-    #         if t < end_times[r] and (r in c[0] or r in c[1]) and c[2] < 0:
-    #             print(c)
-    # if global_collision_counter > 50000 and t > max([t_e for r, t_e in end_times.items()]):
-    #     colls = env.C.getCollisions()
-    #     for c in colls:
-    #         if c[2] < 0:
-    #             print(c)
-    #     print(t)
-    #     env.C.view(False)
-
-    # print('col')
-
     return False
-
 
 # @profile # run with kernprof -l examples/run_planner.py [your environment] [your flags]
 def edge_collision_free_with_moving_obs(
@@ -520,39 +546,28 @@ def edge_collision_free_with_moving_obs(
     end_times,
     resolution=0.1,
 ):
+    """
+    Collision check of a straight-line path for a set of robots between a start and end configuration and time
+    """
     logger.debug(f"edge check for {robots}")
     logger.debug(f"start/end time: {ts}, {te}")
 
-    if te < ts:
+    if te < ts: # Handles edges going backwards in time (for bidirectional planning)
         te, ts = ts * 1.0, te * 1.0
         qs, qe = copy.deepcopy(qe), copy.deepcopy(qs)
 
-    # # compute discretizatoin step
-    N = config_dist(qe, qs) / resolution
+    # Step 1: Discretize the edge into N checkpoints
+    N = config_dist(qe, qs) / resolution # A check for every resolution unit of distance in config space
     tdiff = te - ts
-    N = max(int(tdiff / 1), N)
-    N = max(int(N), 10)
+    N = max(int(tdiff / 1), N) # At least one check per time unit
+    N = max(int(N), 10) # Minimum 10 checkpoints
 
-    # logger.info(f"{N}, {config_dist(qe, qs) / resolution}")
-
-    # print(N)
-
-    # N = sum(part_indices)
-
-    # rel = False
-
-    # if ts > 74 and ts < 75 and te > 84 and te < 85:
-    #     print("AAAAAAAAA")
-    #     print("AAAAAAAAA")
-    #     print("AAAAAAAAA")
-
-    #     print(N)
-
-    #     rel = True
-
+    # Step 2: Generate uniformly spaced time points from ts to te
     times = [ts + tdiff * idx / (N - 1) for idx in range(N)]
-
-    # in addition to the things above, we check the times at which we might start planning again
+        
+    # Step 3: Critical time points (when robot finishes tasks -> mode transitions happen)
+    # Why? Crucial because a mode transition (e.g., robot picks up object) can change 
+    # the collision geometry (discontinuously). Need to check exactly then
     non_escape_end_times = prev_plans.get_non_escape_end_times(env.robots)
     for r, additional_time_to_check in non_escape_end_times.items():
         if ts < additional_time_to_check < te or te < additional_time_to_check < ts:
@@ -564,20 +579,29 @@ def edge_collision_free_with_moving_obs(
 
     start_interpolation_at_index = {}
 
+    # Step 4: Determine the interpolation strategy 
+    # At which time index we should start interpolating toward qe?
     for r in robots:
         if ts < end_times[r] and te > end_times[r]:
+            # CASE 1: Robot r finishes DURING this edge
+            # Start interpolating from where the robot finished toward qe
             for idx in indices:
                 t = times[idx]
                 if t <= end_times[r]:
                     start_interpolation_at_index[r] = idx
 
         elif end_times[r] <= ts:
+            # CASE 2: robot r finishes BEFORE this edge started
+            # Start interpolating from the beginning of the new planned edge toward qe
             start_interpolation_at_index[r] = 0
         elif end_times[r] > te:
+            # CASE 3: robot r finishes AFTER this edge is started 
+            # Don't interpolate, simply follow previous plan for all time samples
             start_interpolation_at_index[r] = indices[-1]
         else:
             start_interpolation_at_index[r] = None
 
+    # Step 5: Store start/end configurations
     q0s = {}
     q1s = {}
     qdiff = {}
@@ -587,37 +611,43 @@ def edge_collision_free_with_moving_obs(
         q1s[i] = qe[i]
         qdiff[i] = qe[i] - qs[i]
 
+    # Step 6: Check middle points first for early collision detection
+    # Instead of checking [0, 1, 2... 10] we check [5, 2, 8, 1, ...]
     indices = generate_binary_search_indices(len(times))
 
+    # Identify other robots (robots: being planned, other_robots: moving obstacles)
     other_robots = []
     for r in env.robots:
         if r not in robots:
             other_robots.append(r)
 
+    # Buffer to build the full configuration of all robots for the env's check
     q_buffer = env.start_pos.state() * 1.0
-
     all_poses = [None] * len(times)
 
+    # Step 8-9: Main collision check loop  
     for idx in indices:
         ql = []
         t = times[idx]
 
-        default_interp = (t - ts) / tdiff
+        default_interp = (t - ts) / tdiff # Interpolation param [0,1]
 
+        # Step 9A: Compute each robot position at time t 
         for i, r in enumerate(robots):
-            # robot_poses[r] = []
+            # CASE 1: Robot hasn't finished (normal case) -> interpolate from qs to qe
             if start_interpolation_at_index[r] == 0:
                 q0 = q0s[i]
-                # q1 = q1s[i]
-                # qdiff = q1 - q0
-
                 p = q0 + qdiff[i] * default_interp
-                # robot_poses[r].append(p)
+            # CASE 2: Robot finished during or before the edge    
             else:
                 logger.debug(f"interpolating {r}")
                 if start_interpolation_at_index[r] >= idx:
+                    # Before robot finished -> use its previous plan
+                    # Robot was free from the start -> simple linear interpolation 
                     p = prev_plans.get_robot_poses_at_time([r], t)[0] * 1.0
                 else:
+                    # Robot finished -> interpolate from finish pose to final edge pose
+                    # Robot has a mixed strategy
                     robot_start_interp_pose = prev_plans.get_robot_poses_at_time(
                         [r], times[start_interpolation_at_index[r]]
                     )[0]
@@ -635,8 +665,8 @@ def edge_collision_free_with_moving_obs(
 
             ql.append(p)
 
-        q = np.concatenate(ql)
-        # all_poses.append(q)
+        # Step 10: Collision check
+        q = np.concatenate(ql) # Combine all robot poses into one config
         all_poses[idx] = q
 
         if not collision_free_with_moving_obs(
@@ -649,13 +679,10 @@ def edge_collision_free_with_moving_obs(
             robots,
             other_robots,
         ):
-            # if rel:
-            #     env.show(True)
-            return False
-        # if rel:
-        #     env.show(True)
-    
-    # TODO: this is an ugly hack and should be done differently
+            return False # Collision found
+
+    # Final sanity check: if generated path doesn't have large discontinuous jumps
+    # TODO: this is an ugly hack and should be done differently (Valentin)
     if len(robots) > 1:
         for i in range(len(all_poses)-1):
             if config_dist(qs.from_flat(all_poses[i]), qs.from_flat(all_poses[i+1])) > resolution * 2:
@@ -663,7 +690,33 @@ def edge_collision_free_with_moving_obs(
 
     return True
 
+"""
+NOTES (Liam):
+Remove later also comments within code, keep for now for better understanding
+=============================================================================================
+PLANNING FUNCTIONS
 
+plan_in_time_space:
+- Unidirectional RRT in joint config-time space. Grows single tree from start, samples random (q,t)
+  nodes, steers under velocity constraints, checks collisions against static env + dyn obstacles 
+  (other_paths), and returns TimedPath in joint space.
+
+plan_in_time_space_bidirectional:
+- Bidirectional RRT in joint config-time space. Grows forward tree from start and reverse tree from
+  sampled goals, alternating expansion and attempting to connect them. Same velocity/collision checks.
+  Generally faster than unidirectional. Returns TimedPath in joint space. 
+
+shortcut_with_dynamic_obstacles:
+- Post-processing optimizer. Discretizes RRT path, then repeatedly picks two random waypoints and 
+  tries to replace the segment between them with straight-line interpolation. Accepts only if 
+  collision-free against static env + dyn obstacles. Smooths and shortens the path.
+
+plan_robots_in_dyn_env:
+- Wrapper functions managing the planning for one task, choosing appropriate planner, optionally
+  post-processes with shortcutting, splits the planned joint path into per-robot paths, and returns
+  per-robot separated paths and final joint configuration  
+=============================================================================================
+"""
 def plan_in_time_space(
     ptc,
     env: BaseProblem,
@@ -675,10 +728,13 @@ def plan_in_time_space(
     t_lb,
     v_max=0.5,
 ) -> TimedPath:
+    """
+    Unidirectional RRT planner operating in joint space-time configuration
+    Plans for a specific subset of robots (active) while treating other robots as dynamic obstacles
+    """
+    # Setup and initialization
     computation_start_time = time.time()
-
     max_iter = 50000
-
     conf_type = type(env.get_start_pos())
 
     logger.info("start_time", t0)
@@ -690,6 +746,7 @@ def plan_in_time_space(
 
     logger.info("start state", q0.state())
 
+    # Tree initialization
     tree = Tree(Node(t0, q0))
     tree.prev_plans = prev_plans
     tree.robots = robots
@@ -702,51 +759,28 @@ def plan_in_time_space(
             assert False
 
         t_diff = rnd_node.t - close_node.t
-        q_diff = rnd_node.q.state() - close_node.q.state()
-        length = config_dist(rnd_node.q, close_node.q)
+        q_diff = rnd_node.q.state() - close_node.q.state() # Vector form (direction + magnitude)
+        length = config_dist(rnd_node.q, close_node.q) # Scalar distance
 
-        v = length / t_diff
-
-        # if length < 1e-3:
-        #     return None, None
+        v = length / t_diff # Required velocity
 
         logger.debug(f"length {length}")
 
         if v > v_max:
-            return None, None
-            assert False
+            return None, None # Physically infeasible (enforces kinodynamic constraints)
 
-        # max_q = 1
-
-        # if length < max_q:
-        #     return rnd_node.t, rnd_node.q
-
-        # t_req = 1/length * t_diff
-        # t = close_node.t + t_req
-        # q = close_node.q.state() + t_req * v * q_diff / length
-
+        # Either, can reach sample directly
         if t_diff < max_stepsize:
-            # print('reached', rnd_node.q.state())
-            return rnd_node.t, rnd_node.q
+            return rnd_node.t, rnd_node.q 
 
-        # t_m = min(max_q / v_max, max_stepsize)
+        # Otherwise, take a step of size max_stepsize toward sample
         t_m = min(max_stepsize * 1.0, t_diff)
 
         if length < 1e-3:
             return None, None
-        # # print('not')
 
         t = close_node.t + t_m
-        q = close_node.q.state() + t_m * v * q_diff / length
-
-        # print('v', v)
-        # print('scale', max_stepsize / length)
-        # print('close in steer', close_node.q.state())
-        # print('rnd_node in steer', rnd_node.q.state())
-        # print('q_diff', q_diff)
-        # print('q in steer', q)
-        # if q[0] > 1:
-        #     raise ValueError
+        q = close_node.q.state() + t_m * v * q_diff / length # t_m*v (displacement magnitude), q_diff/length (unit direction vector)
 
         q_list = []
         offset = 0
@@ -773,19 +807,19 @@ def plan_in_time_space(
         return t_rnd, conf_type.from_list(q_goal_as_list)
 
     def sample_uniform(t_ub, goal_sampling_probability=0.1):
+        """Informed sampling: narrow the sampling region to a box around start-goal corridor"""
         informed_sampling = True
         if informed_sampling:
-            # sample from box
+            # Compute midpoint between q0 and qg
             q0_state = q0.state()
             qg_state = sampled_goals[0][1].state()
-
             mid = (q0_state + qg_state) / 2
 
+            # Compute box half-width
             max_goal_time = max([g[0] for g in sampled_goals])
+            c_max = (max_goal_time - t0) * v_max # Max distance robot can travel in available time
 
-            c_max = (max_goal_time - t0) * v_max
-            # c_min = config_dist(q0, sampled_goals[0][1])
-
+            # Sample in box: [mid-c_max/2, mid+c_max/2]
             s = []
             i = 0
             for r in robots:
@@ -799,22 +833,23 @@ def plan_in_time_space(
                     lo = max(lims[0, j], lo)
                     hi = min(lims[1, j], hi)
 
+                    # For each DOF sample uniformly within box
                     s.append(np.random.uniform(lo, hi))
 
                     i += 1
 
+            # Compute feasible time window for this sample
             conf = q0.from_flat(np.array(s))
-
             min_dt_from_start = config_dist(q0, conf) / v_max
-
-            min_t_sample = t0 + min_dt_from_start
-            max_t_sample = (
+            min_t_sample = t0 + min_dt_from_start # Earliest robot can reach this config
+            max_t_sample = ( # Latest time s.t. robot can still reach the goal
                 max_goal_time - config_dist(conf, sampled_goals[0][1]) / v_max
             )
 
             if min_t_sample > max_t_sample:
                 return max_goal_time + 1, q0.from_flat(np.random.rand(len(q0.state())))
 
+            # Sample time uniformly within bounds
             t_rnd = np.random.uniform(min_t_sample, max_t_sample)
         else:
             max_goal_time = max([g[0] for g in sampled_goals])
@@ -825,18 +860,10 @@ def plan_in_time_space(
                 raise ValueError
 
             q_rnd = []
-            # rnd_config = env.sample_config_uniform_in_limits()
-            # print(rnd_config.state())
-            # for i, r in enumerate(env.robots):
-            #     if r not in robots:
-            #         continue
+    
             for r in robots:
                 idx = env.robot_idx[r]
-
                 lims = env.limits[:, idx]
-
-                # print(lims)
-
                 dim = env.robot_dims[r]
 
                 rnd_uni_0_1 = np.random.rand(dim)
@@ -849,6 +876,7 @@ def plan_in_time_space(
         return t_rnd, conf
 
     def project_sample_to_preplanned_path(t, q):
+        """Enforce that robots follow their previous plan when they're still busy"""
         logger.debug("projecting")
         q_new = q
 
@@ -859,12 +887,13 @@ def plan_in_time_space(
 
         return q_new
 
+    # Time bounds computation
     latest_end_time = max([end_times[r] for r in robots])
-    t_lb = max(latest_end_time + 1, t_lb)
+    t_lb = max(latest_end_time + 1, t_lb) # New plan can only start at t_lb (after all robots are free + buffer)
 
     logger.debug(f"end_time: {t_lb}")
 
-    # estimate distance
+    # Estimate start-goal distance
     start_poses = prev_plans.get_robot_poses_at_time(robots, t_lb)
     goal_pose = goal.sample(None)
     goal_config = []
@@ -873,22 +902,24 @@ def plan_in_time_space(
         dim = env.robot_dims[r]
         goal_config.append(goal_pose[offset : offset + dim])
         offset += dim
+
+    # Configuration space distance between start and goal
     d = config_dist(conf_type.from_list(goal_config), conf_type.from_list(start_poses))
 
     logger.debug("Goal pose", conf_type.from_list(goal_config).state())
     logger.debug("start/goal dist", d)
 
-    # compute max time from it
-    max_t = t_lb + 1 + (d / v_max) * 10
+    # Compute time bounds (with calculated config space distance and max velocity)
+    max_t = t_lb + 1 + (d / v_max) * 10 # Conservative upper bound (allows for detours around obstacles)
 
     escape_path_end_time = prev_plans.get_final_time()
-    max_t = max(max_t, escape_path_end_time)
+    # Needs to span at least as long as any existing plan to ensure correct collision checking
+    max_t = max(max_t, escape_path_end_time) # Consider other robot's escape paths
 
     logger.debug("start_times", end_times)
     logger.debug("max time", max_t)
 
-    curr_t_ub = max([end_times[r] for r in robots]) + (d / v_max) * 3
-
+    curr_t_ub = max([end_times[r] for r in robots]) + (d / v_max) * 3 # Start with more reasonable initial bound
     curr_t_ub = max(curr_t_ub, t_lb)
 
     logger.debug("times for lb ", t0, d / v_max)
@@ -898,42 +929,17 @@ def plan_in_time_space(
     sampled_times = []
     sampled_pts = []
 
-    other_robots = []
+    other_robots = [] # All robots not being planned -> dynamic obstacles
     for r in env.robots:
         if r not in robots:
             other_robots.append(r)
 
-    # print("Goal pose")
-    # res = collision_free_with_moving_obs(
-    #     env,
-    #     t_lb,
-    #     conf_type.from_list(goal_config).state(),
-    #     env.start_pos.state() * 1.0,
-    #     prev_plans,
-    #     end_times,
-    #     robots,
-    #     other_robots,
-    #     robot_joints,)
-    # print(res)
-    # env.C.view(True)
-
-    # if not collision_free_with_moving_obs(
-    #     env,
-    #     t0,
-    #     q0.state(),
-    #     env.start_pos.state() * 1.0,
-    #     prev_plans,
-    #     end_times,
-    #     robots,
-    #     other_robots,
-    # ):
-    #     print("Start pose not feasible")
-    #     env.C.view(True)
-    #     assert False
-
+    # MAIN PLANNING LOOP 
     iter = 0
     while True:
         iter += 1
+
+        # Termination checks
         if ptc.should_terminate(0, time.time() - computation_start_time):
             break
 
@@ -941,16 +947,17 @@ def plan_in_time_space(
             print("Max iter reached, stopping")
             break
 
-        # increase upper bound that we are sampling
+        # Gradually expand time horizon (upper bound that we are sampling)
         if iter % 50:
             curr_t_ub += 1
             curr_t_ub = min(curr_t_ub, max_t)
 
-        # sample pt
+        # Step 1: sampling phase
         # sample time and position
         rnd = random.random()
+        goal_sampling_probability = 0.1 # Balance between exploration/exploitation
 
-        goal_sampling_probability = 0.1
+        # First always samples goal because need at least one goal to define informed sampling region  
         if len(sampled_goals) == 0 or rnd < goal_sampling_probability:
             t_rnd, q_sampled = sample_goal(curr_t_ub)
             sampled_goals.append((t_rnd, q_sampled))
@@ -958,42 +965,46 @@ def plan_in_time_space(
         else:
             t_rnd, q_sampled = sample_uniform(curr_t_ub)
 
+        # Projection after sampling
+        # Why? sample_uniform doesn't know about multi-robot constraints (e.g., prev_plans), if robot 
+        # still executing -> override sample with exact pose at time t along prev_path (through interpolation) 
+        # Simpler to sample and project, than baking constraint into sampling itself 
         q_rnd = project_sample_to_preplanned_path(t_rnd, q_sampled)
 
         logger.debug("rnd state,", q_rnd.state())
 
-        # check if there is a chance that we can reach the goal (or the start)
+        # Step 2: feasibility checks
+        # Checks are ordered by computational cost (cheap to expensive)
+        # Check 1: Reachable from start?
         time_from_start = t_rnd - t0
         d_from_start = config_dist(q0, q_rnd, "max")
         if d_from_start / time_from_start > v_max:
-            # print("pos not reachable from start")
             continue
-
+        
+        # Check 2: Can reach some goal?
         reachable_goal = False
         for tg, qg in sampled_goals:
             time_from_goal = tg - t_rnd
-            # we do not need to plan to a time which does not yet have a goal that we can reach.
+            # We do not need to plan to a time which does not yet have a goal that we can reach.
             if tg < t_rnd:
-                continue
+                continue # Skip goals from past (can't go backward in time)
 
             if np.linalg.norm(qg.state() - q_rnd.state()) < 1e-3:
-                reachable_goal = True
+                reachable_goal = True # Already at goal
                 break
-
+            
+            # Use max norm -> conservative, uses worst-case velocity
             d_from_goal = config_dist(qg, q_rnd, "max")
             if d_from_goal / time_from_goal <= v_max:
                 reachable_goal = True
                 break
 
-            # print("goal dists", d_from_goal, time_from_goal, d_from_goal/time_from_goal)
-
         if not reachable_goal:
             logger.debug("No reachable goal for the sampled node.")
             logger.debug("times", tg, t_rnd)
-
             continue
 
-        # check if sample is valid
+        # Check 3: Collision-free (valid) sample?
         if not collision_free_with_moving_obs(
             env,
             t_rnd,
@@ -1009,81 +1020,20 @@ def plan_in_time_space(
         if t_rnd < t0:
             raise ValueError
 
-        # find closest pt in tree
+        # Step 3: tree extension
+        # Find nearest neighbour (closest pt in existing tree)
         n_close = tree.get_nearest_neighbor(Node(t_rnd, q_rnd), v_max)
 
-        # if len(tree.nodes) > 200 and loopy_bois % 5 == 0:
-        #     plt.figure()
-        #     for n in tree.nodes:
-        #         if n.parent is None:
-        #             continue
-
-        #         x0 = n.parent.q.state()[0]
-        #         y0 = n.parent.q.state()[1]
-
-        #         x1 = n.q.state()[0]
-        #         y1 = n.q.state()[1]
-
-        #         # print(x0, x1, y0, y1)
-
-        #         plt.plot([x0, x1], [y0, y1])
-
-        #     plt.figure()
-        #     for n in tree.nodes:
-        #         if n.parent is None:
-        #             continue
-
-        #         x0 = n.parent.q.state()[3]
-        #         y0 = n.parent.q.state()[4]
-
-        #         x1 = n.q.state()[3]
-        #         y1 = n.q.state()[4]
-
-        #         # print(x0, x1, y0, y1)
-
-        #         plt.plot([x0, x1], [y0, y1])
-
-        #     plt.show()
-
-        # if len(q_rnd.state()) > 3:
-        #     rnd_nodes = []
-        #     for _ in range(5000):
-        #         t_rnd, q_uni_rnd = uniform_sample(curr_t_ub)
-        #         q_rnd = project_sample_to_preplanned_path(t_rnd, q_uni_rnd)
-
-        #         if collision_free_with_moving_obs(t_rnd, q_rnd.state()):
-        #             rnd_nodes.append(Node(t_rnd, q_rnd))
-
-        #     fig = plt.figure(figsize=(12, 12))
-        #     ax = fig.add_subplot(projection='3d')
-
-        #     for i, n in enumerate(rnd_nodes):
-        #         k = 10
-        #         node_dists = tree.batch_dist_fun(n, rnd_nodes, v_max)
-        #         ind = np.argpartition(node_dists, k)[:k]
-
-        #         for j in ind[1:]:
-        #             # print(n.t, rnd_nodes[j].t)
-        #             if abs(n.t - rnd_nodes[j].t) < 1e-3:
-        #                 continue
-        #             if not np.isinf(node_dists[j]) and edge_collision_free_with_moving_obs(rnd_nodes[j].q, n.q, rnd_nodes[j].t, n.t):
-        #                 ax.plot([n.q.state()[0], rnd_nodes[j].q.state()[0]], [n.q.state()[1], rnd_nodes[j].q.state()[1]], [n.t, rnd_nodes[j].t], color='black')
-
-        #     ax.set_xlim([-2, 2])
-        #     ax.set_ylim([-2, 2])
-
-        #     ax.set_xlabel('x')
-
-        #     plt.show()
-
         if n_close is None:
-            continue
+            continue # No feasible connection exists
 
         added_pt = False
         q_new = None
         t_new = None
-
         extend = False
+
+        # Takes multiple small steps toward the sample
+        # Greedily extends the tree as far as possible toward the sample in one iteration
         if extend:
             t_goal = t_rnd
             q_goal = q_rnd
@@ -1093,12 +1043,16 @@ def plan_in_time_space(
             while True:
                 steps += 1
 
-                t_next, q_next = steer(n_prev, Node(t_goal, q_goal), 5)
+                t_next, q_next = steer(n_prev, Node(t_goal, q_goal), 5) # Steer small step (5) from n_prev to sample
                 if t_next is None or q_next is None:
-                    break
+                    break # Stop if steer fails
+                
+                # Second projection on steered sample
+                q_next = project_sample_to_preplanned_path(t_next, q_next) # Busy robots follow their plans
 
-                q_next = project_sample_to_preplanned_path(t_next, q_next)
-
+                # Collision check (point and edge)
+                # Why check point if after we either way check edge?
+                # If endpoint collides, no reason to then check whole edge (makes sense here  with many small steps)
                 if not collision_free_with_moving_obs(
                     env,
                     t_next,
@@ -1122,7 +1076,7 @@ def plan_in_time_space(
                     end_times,
                     resolution=env.collision_resolution,
                 ):
-                    # add to tree
+                    # Add node to tree
                     tree.add_node(Node(t_next, q_next), n_prev)
 
                     n_prev = tree.nodes[-1]
@@ -1132,19 +1086,21 @@ def plan_in_time_space(
                     q_new = q_next
 
                     if np.linalg.norm(q_goal.state() - q_new.state()) < 1e-3:
-                        break
+                        break # Stop when reaching sample 
                 else:
-                    break
+                    break # Stop when hitting a collision
+        # Take one big step toward the sample
         else:
-            # steer towards pt
-            t_new, q_new = steer(n_close, Node(t_rnd, q_rnd), max_stepsize=30)
+            t_new, q_new = steer(n_close, Node(t_rnd, q_rnd), max_stepsize=30) # Steer one big step
 
+            # Same 
             if t_new is None or q_new is None:
                 continue
+            
+            # Second projection (on steered sample)
+            q_new = project_sample_to_preplanned_path(t_new, q_new) 
 
-            q_new = project_sample_to_preplanned_path(t_new, q_new)
-
-            # check if edge is collision-free
+            # Collision check (edge only)
             if edge_collision_free_with_moving_obs(
                 env,
                 n_close.q,
@@ -1156,115 +1112,49 @@ def plan_in_time_space(
                 end_times,
                 resolution=env.collision_resolution,
             ):
-                # add to tree
+                # Add node to tree
                 tree.add_node(Node(t_new, q_new), n_close)
 
                 sampled_times.append(t_new)
                 sampled_pts.append(q_new.state())
 
-                # if cnt % 100 == 0 and cnt > 0:
-                #     plt.figure()
-                #     plt.hist(sampled_times, bins=100)
-
-                #     plt.figure()
-                #     plt.scatter([s[0] for s in sampled_pts], [s[1] for s in sampled_pts])
-
-                #     plt.figure()
-                #     plt.scatter([s[3] for s in sampled_pts], [s[4] for s in sampled_pts])
-
-                #     plt.show()
-
                 added_pt = True
                 logger.debug(f"succ at time {t_new}")
             else:
-                # plt.plot([n_close.q.state()[0], q_new.state()[0]], [n_close.q.state()[1], q_new.state()[1]])
                 logger.debug(f"failed at time {t_new}")
                 logger.debug(f"Tree size {len(tree.nodes)}")
 
-                # env.C.view(False)
-
-        if (
-            added_pt
-            and goal.satisfies_constraints(q_new.state(), mode=None, tolerance=1e-5)
-            and t_new > t_lb
+        # Step 4: goal check 
+        # Condition 1: successfully added to tree
+        # Condition 2: in goal region
+        # Condition 3: finished after earliest allowed time
+        if (added_pt and 
+            goal.satisfies_constraints(q_new.state(), mode=None, tolerance=1e-5) and 
+            t_new > t_lb
         ):
-            # plt.show()
+            # Extract path by backtracking through parents (gives goal -> start)
             configurations = [q_new.state()]
             times = [t_new]
-
             p = n_close
 
             while p.parent is not None:
                 configurations.append(p.q.state())
                 times.append(p.t)
-
                 p = p.parent
 
-            configurations.append(p.q.state())
+            configurations.append(p.q.state()) # Add root
             times.append(p.t)
-
-            # print(times[::-1])
-
-            # print(configurations)
-            # print(times)
 
             computation_end_time = time.time()
             logger.debug(f"Took {computation_end_time - computation_start_time}s")
 
-            # for k in range(len(robots)):
-            #     # plt.figure()
-            #     # for n in tree.nodes:
-            #     #     if n.parent is None:
-            #     #         continue
-
-            #     #     x0 = n.parent.q.state()[0 + k*2]
-            #     #     y0 = n.parent.q.state()[1 + k*2]
-
-            #     #     x1 = n.q.state()[0 + k*2]
-            #     #     y1 = n.q.state()[1 + k*2]
-
-            #     #     # print(x0, x1, y0, y1)
-
-            #     #     plt.plot([x0, x1], [y0, y1])
-            #     fig = plt.figure(figsize=(10, 8)) # Create a new figure for each robot
-            #     # Add a 3D subplot to the figure
-            #     ax = fig.add_subplot(111, projection='3d')
-            #     ax.set_title(f"Robot {k} Path in 3D (t-component as Z)")
-            #     ax.set_xlabel("X-coordinate")
-            #     ax.set_ylabel("Y-coordinate")
-            #     ax.set_zlabel("T-component")
-
-            #     # Iterate through the nodes in the tree to plot the path segments
-            #     for n in tree.nodes:
-            #         if n.parent is None:
-            #             continue # Skip the root node as it has no parent to draw a line from
-
-            #         # Parent's coordinates and t-component
-            #         dim = 3
-            #         x0 = n.parent.q.state()[0 + k*dim] # X-coordinate of parent
-            #         y0 = n.parent.q.state()[1 + k*dim] # Y-coordinate of parent
-            #         z0 = n.parent.t             # T-component of parent (now Z)
-
-            #         # Current node's coordinates and t-component
-            #         x1 = n.q.state()[0 + k*dim] # X-coordinate of current node
-            #         y1 = n.q.state()[1 + k*dim] # Y-coordinate of current node
-            #         z1 = n.t             # T-component of current node (now Z)
-
-            #         # Plot the line segment in 3D
-            #         # The 'k*2' from your original code was likely for indexing into a flattened state
-            #         # array that contained multiple robot states. Here, we assume n.q.state()
-            #         # directly gives the x, y for the current robot, and n.t is its time.
-            #         ax.plot([x0, x1], [y0, y1], [z0, z1], marker='o', linestyle='-')
-
-            # plt.show()
-
+            # Reverse to get (start -> goal)
             return TimedPath(time=times[::-1], path=configurations[::-1])
 
     if configurations is None:
         return None
 
     return None
-
 
 def plan_in_time_space_bidirectional(
     ptc,
@@ -1277,20 +1167,21 @@ def plan_in_time_space_bidirectional(
     t_lb,
     v_max=0.5,
 ) -> TimedPath:
+    """
+    Bidirectional RRT planner operating in joint space-time configuration
+    Plans for a specific subset of robots (active) while treating other robots as dynamic obstacles
+    """
     computation_start_time = time.time()
     conf_type = type(env.get_start_pos())
-
-    # print("start_time", t0)
-    # print("robots", robots)
-    # print("earliest end time", t_lb)
 
     start_configuration = prev_plans.get_robot_poses_at_time(robots, t0)
     q0 = conf_type.from_list(start_configuration)
 
     logger.info(f"start state {q0.state()}")
 
+    # Tree initialization (forward and reverse)
     t_fwd = Tree(Node(t0, q0))
-    t_rev = Tree(None, reverse=True)
+    t_rev = Tree(None, reverse=True) # Empty because don't know goal configuration yet
 
     t_fwd.prev_plans = prev_plans
     t_fwd.robots = robots
@@ -1310,40 +1201,27 @@ def plan_in_time_space_bidirectional(
             assert False
 
         t_diff = rnd_node.t - close_node.t
-        q_diff = rnd_node.q.state() - close_node.q.state()
-        length = config_dist(rnd_node.q, close_node.q)
+        q_diff = rnd_node.q.state() - close_node.q.state() # Vector form (direction + magnitude)
+        length = config_dist(rnd_node.q, close_node.q) # Scalar distance
 
-        v = length / t_diff
-
-        # if length < 1e-3:
-        #     return None, None
+        v = length / t_diff # Required velocity
 
         if v > v_max:
-            return None, None
+            return None, None # Physically infeasible (enforces kinodynamic constraints)
             assert False
 
-        # max_q = 1
-
-        # if length < max_q:
-        #     return rnd_node.t, rnd_node.q
-
-        # t_req = 1/length * t_diff
-        # t = close_node.t + t_req
-        # q = close_node.q.state() + t_req * v * q_diff / length
-
+        # Either, can reach sample directly
         if t_diff < max_stepsize:
-            # print('reached', rnd_node.q.state())
             return rnd_node.t, rnd_node.q
 
-        # t_m = min(max_q / v_max, max_stepsize)
+        # Otherwise, take a step of size max_stepsize toward sample
         t_m = min(max_stepsize * 1.0, t_diff)
 
         if length < 1e-3:
             return None, None
-        # # print('not')
 
         t = close_node.t + t_m
-        q = close_node.q.state() + t_m * v * q_diff / length
+        q = close_node.q.state() + t_m * v * q_diff / length # t_m*v -> displacement magnitude, q_diff/length -> unit direction vector
 
         q_list = []
         offset = 0
@@ -1354,33 +1232,35 @@ def plan_in_time_space_bidirectional(
 
         return t, conf_type.from_list(q_list)
 
-    # we still steer from close to rnd.
+    # We still steer from close to rnd.
     # But we now assume that close node has a higher time.
     def reverse_steer(close_node: Node, rnd_node: Node, max_stepsize=30):
-        if close_node.t < rnd_node.t:
+        if close_node.t < rnd_node.t: # DIFFERENCE IS HERE!
             logger.warn("close", close_node.t)
             logger.warn("goal", rnd_node.t)
             assert False
 
         t_diff = close_node.t - rnd_node.t
-        q_diff = rnd_node.q.state() - close_node.q.state()
-        length = config_dist(rnd_node.q, close_node.q)
+        q_diff = rnd_node.q.state() - close_node.q.state() # Vector form (direction + magnitude)
+        length = config_dist(rnd_node.q, close_node.q) # Scalar distance
 
-        v = length / t_diff
+        v = length / t_diff # Required velocity
 
         if v > v_max:
-            return None, None
+            return None, None # Physically infeasible (enforces kinodynamic constraints)
 
+        # Either, can reach sample directly
         if t_diff < max_stepsize:
             return rnd_node.t, rnd_node.q
 
+        # Otherwise, take a step of size max_stepsize toward sample
         t_m = min(max_stepsize * 1.0, t_diff)
 
         if length < 1e-3:
             return None, None
 
-        t = close_node.t - t_m
-        q = close_node.q.state() + t_m * v * q_diff / length
+        t = close_node.t - t_m # DIFFERENCE IS HERE!
+        q = close_node.q.state() + t_m * v * q_diff / length # t_m*v -> displacement magnitude, q_diff/length -> unit direction vector
 
         q_list = []
         offset = 0
@@ -1405,19 +1285,19 @@ def plan_in_time_space_bidirectional(
         return t_rnd, conf_type.from_list(q_goal_as_list)
 
     def sample_uniform():
+        """Informed sampling: narrow the sampling region to a box around start-goal corridor"""
         informed_sampling = True
         if informed_sampling:
-            # sample from box
+            # Compute midpoint between q0 and qg
             q0_state = q0.state()
             qg_state = sampled_goals[0][1].state()
-
             mid = (q0_state + qg_state) / 2
 
+            # Compute box half-width
             max_goal_time = max([g[0] for g in sampled_goals])
-
             c_max = (max_goal_time - t0) * v_max
-            # c_min = config_dist(q0, sampled_goals[0][1])
 
+            # Sample in box: [mid-c_max/2, mid+c_max/2]
             s = []
             i = 0
             for r in robots:
@@ -1431,22 +1311,22 @@ def plan_in_time_space_bidirectional(
                     lo = max(lims[0, j], lo)
                     hi = min(lims[1, j], hi)
 
+                    # For each DOF sample uniformly within box
                     s.append(np.random.uniform(lo, hi))
 
                     i += 1
 
+            # Compute feasible time window for this sample
             conf = q0.from_flat(np.array(s))
-
             min_dt_from_start = config_dist(q0, conf) / v_max
-
-            min_t_sample = t0 + min_dt_from_start
-            max_t_sample = (
+            min_t_sample = t0 + min_dt_from_start # Earliest robot can reach this config
+            max_t_sample = ( # Latest time s.t. robot can still reach the goal
                 max_goal_time - config_dist(conf, sampled_goals[0][1]) / v_max
             )
 
             if min_t_sample > max_t_sample:
                 return max_goal_time + 1, q0.from_flat(np.random.rand(len(q0.state())))
-
+            # Sample time uniformly within bounds
             t_rnd = np.random.uniform(min_t_sample, max_t_sample)
         else:
             max_goal_time = max([g[0] for g in sampled_goals])
@@ -1457,38 +1337,23 @@ def plan_in_time_space_bidirectional(
                 raise ValueError
 
             q_rnd = []
-            # rnd_config = env.sample_config_uniform_in_limits()
-            # print(rnd_config.state())
-            # for i, r in enumerate(env.robots):
-            #     if r not in robots:
-            #         continue
+   
             for r in robots:
                 idx = env.robot_idx[r]
-
                 lims = env.limits[:, idx]
-
-                # print(lims)
-
                 dim = env.robot_dims[r]
 
                 rnd_uni_0_1 = np.random.rand(dim)
                 q = rnd_uni_0_1 * (lims[1, :] - lims[0, :]) + lims[0, :]
 
-                # print('rnd val', rnd_uni_0_1)
-                # print((lims[1, :] - lims[0, :]))
-                # print(lims[0, :])
-                # print('q', q)
-
                 q_rnd.append(q * 1.0)
-                # q_rnd.append(rnd_config[i])
 
             conf = conf_type.from_list(q_rnd)
-
-            # print(q_rnd)
 
         return t_rnd, conf
 
     def project_sample_to_preplanned_path(t: float, q: Configuration) -> Configuration:
+        """Enforce that robots follow their previous plan when they're still busy"""
         q_new = q
 
         for i, r in enumerate(robots):
@@ -1498,12 +1363,13 @@ def plan_in_time_space_bidirectional(
 
         return q_new
 
+    # Time bounds computation
     latest_end_time = max([end_times[r] for r in robots])
-    t_lb = max(latest_end_time + 1, t_lb)
+    t_lb = max(latest_end_time + 1, t_lb) # New plan can only start at t_lb (after all robots are free + buffer)
 
     logger.info(f"end_time: {t_lb}")
 
-    # estimate distance
+    # Estimate start-goal distance
     start_poses = prev_plans.get_robot_poses_at_time(robots, t_lb)
     goal_pose = goal.sample(None)
     goal_config = []
@@ -1512,55 +1378,43 @@ def plan_in_time_space_bidirectional(
         dim = env.robot_dims[r]
         goal_config.append(goal_pose[offset : offset + dim])
         offset += dim
+    # Configuration space distance between start and goal
     d = config_dist(conf_type.from_list(goal_config), conf_type.from_list(start_poses))
 
     logger.debug(f"Goal pose {conf_type.from_list(goal_config).state()}")
     logger.debug(f"start/goal dist {d}")
 
-    # compute max time from it
-    max_t = t_lb + 1 + (d / v_max) * 10
+    # Compute time bounds (with calculated config space distance and max velocity)
+    max_t = t_lb + 1 + (d / v_max) * 10 # Conservative upper bound (allows for detours around obstacles)
 
     escape_path_end_time = prev_plans.get_final_time()
-    max_t = max(max_t, escape_path_end_time)
+    # Needs to span at least as long as any existing plan to ensure correct collision checking
+    max_t = max(max_t, escape_path_end_time) # Consider other robot's escape paths
 
     logger.debug(f"start_times{end_times}")
     logger.debug(f"max time{max_t}")
 
-    # curr_t_ub = t_lb + (d / v_max) * 3
-    curr_t_ub = max([end_times[r] for r in robots]) + (d / v_max) * 3
-    # curr_t_ub = max_t
-
+    curr_t_ub = max([end_times[r] for r in robots]) + (d / v_max) * 3 # Start with more reasonable initial bound
     curr_t_ub = max(curr_t_ub, t_lb)
 
     configurations = None
 
-    other_robots = []
+    other_robots = [] # All robots not being planned -> dynamic obstacles
     for r in env.robots:
         if r not in robots:
             other_robots.append(r)
 
-    # print("Goal pose")
-    # res = collision_free_with_moving_obs(
-    #     env,
-    #     t_lb,
-    #     conf_type.from_list(goal_config).state(),
-    #     env.start_pos.state() * 1.0,
-    #     prev_plans,
-    #     end_times,
-    #     robots,
-    #     other_robots,
-    #     robot_joints,)
-    # print(res)
-    # env.C.view(True)
-
+    # Phase 1: goal sampling
+    # The reverse tree needs at least one collision-free goal node to start growing backward from 
     iter = 0
     while True:
         iter += 1
 
         t_rnd, q_rnd = sample_goal(max_t)
         if iter == 1:
-            t_rnd = max_t
+            t_rnd = max_t # First sample at latest time (conservative upper bound computed earlier)
 
+        # Collision check of goal with env at time t_rnd
         res = collision_free_with_moving_obs(
             env,
             t_rnd,
@@ -1573,16 +1427,17 @@ def plan_in_time_space_bidirectional(
         )
 
         if res:
-            t_rev.add_node(Node(t_rnd, q_rnd), None)
+            t_rev.add_node(Node(t_rnd, q_rnd), None) # Add to reverse tree (root nodes of reverse tree, no parent)
             sampled_goals.append((t_rnd, q_rnd))
 
         if len(sampled_goals) > 0 and iter > 50:
-            break
+            break # Got at least one valid goal
 
         if iter > 10000:
             print("Max iters in goal sampling.")
             return None
 
+    # Start pose validation (check before expensive tree growth)
     if not collision_free_with_moving_obs(
         env,
         t0,
@@ -1594,19 +1449,20 @@ def plan_in_time_space_bidirectional(
         other_robots,
     ):
         logger.warn("Start pose not feasible")
-        # env.C.view(True)
         return None
 
+    # Phase 2: MAIN PLANNING LOOP 
     max_iters = 50000
     for iter in range(max_iters):
         tmp = t_a
         t_a = t_b
-        t_b = tmp
+        t_b = tmp # Swap active and passive trees at every iteration (balanced iterations)
 
+        # Termination checks
         if ptc.should_terminate(0, time.time() - computation_start_time):
             break
 
-        # increase upper bound that we are sampling
+        # Gradually expand time horizon (upper bound that we are sampling)
         if iter % 50:
             curr_t_ub += 1
             curr_t_ub = min(curr_t_ub, max_t)
@@ -1616,40 +1472,43 @@ def plan_in_time_space_bidirectional(
             logger.debug(f"num nodes in trees: {len(t_a.nodes)} {len(t_b.nodes)}")
             logger.debug(f"Current t_ub {curr_t_ub}")
 
+         # Step 1: sampling phase
+         # No goal sampling, already sampled in PHASE 1
         t_rnd, q_sampled = sample_uniform()
+        q_rnd = project_sample_to_preplanned_path(t_rnd, q_sampled) # Only free robot get randomly sampled positions
 
-        q_rnd = project_sample_to_preplanned_path(t_rnd, q_sampled)
-
-        # check if there is a chance that we can reach the goal (or the start)
+        # Step 2: feasibility checks
+        # Checks are ordered by computational cost (cheap to expensive)
+        # Check 1: Reachable from start?
         time_from_start = t_rnd - t0
         d_from_start = config_dist(q0, q_rnd, "max")
         if d_from_start / time_from_start > v_max:
             continue
 
+        # Check 2: Can reach some goal?
         reachable_goal = False
         for tg, qg in sampled_goals:
             time_from_goal = tg - t_rnd
-            # we do not need to plan to a time which does not yet have a goal that we can reach.
+            # We do not need to plan to a time which does not yet have a goal that we can reach.
             if tg < t_rnd:
-                continue
+                continue # Skip goals from past (can't go backward in time)
 
             if np.linalg.norm(qg.state() - q_rnd.state()) < 1e-3:
-                reachable_goal = True
+                reachable_goal = True # Already at goal
                 break
-
+            
+            # Use max norm -> conservative, use worst-case velocity
             d_from_goal = config_dist(qg, q_rnd, "max")
             if d_from_goal / time_from_goal <= v_max:
                 reachable_goal = True
                 break
-
-            # print("goal dists", d_from_goal, time_from_goal, d_from_goal/time_from_goal)
 
         if not reachable_goal:
             logger.warn("No reachable goal for the sampled node.")
             # print("times", tg, t_rnd)
             continue
 
-        # check if sample is valid
+        # Check 3: Collision-free (valid) sample?
         if not collision_free_with_moving_obs(
             env,
             t_rnd,
@@ -1660,37 +1519,32 @@ def plan_in_time_space_bidirectional(
             robots,
             other_robots,
         ):
-            # print('invalid sample')
-            # env.show(True)
-            # if iter >= 1000 and iter % 500 == 0:
-            #     # for some reason there is some bug that seems fixed when displaying a conf once.
-            #     # I do not know what the reason is. I suspect that I am somehow mishandling modes.
-            #     # Not sure how that would be fixed by doing this though.
-            #     # env.show(False)
-            #     if isinstance(env, rai_env):
-            #         env.C.view_recopyMeshes()
-            # env.C.view_close()
             continue
 
         if t_rnd < t0:
             raise ValueError
 
-        # find closest pt in tree
+        # Step 3: tree extension
+        # Find nearest neighbour (closest pt in existing tree)
         n_close = t_a.get_nearest_neighbor(Node(t_rnd, q_rnd), v_max)
 
-        assert n_close is not None
+        assert n_close is not None # Feasible connections should always exists
 
-        # go in dir and check collision
+        # Steering with direction awareness
         if t_a.reverse:
+            # Time decreases (backward tree)
             t_new, q_new = reverse_steer(n_close, Node(t_rnd, q_rnd), max_stepsize=10)
         else:
+            # Time increases (forward tree)
             t_new, q_new = steer(n_close, Node(t_rnd, q_rnd), max_stepsize=10)
 
         if t_new is None or q_new is None:
             continue
-
+        
+        # Project steered sample to enforce robot constraints at this new time 
         q_new = project_sample_to_preplanned_path(t_new, q_new)
 
+        # Collision check 
         if edge_collision_free_with_moving_obs(
             env,
             n_close.q,
@@ -1702,15 +1556,18 @@ def plan_in_time_space_bidirectional(
             end_times,
             resolution=env.collision_resolution,
         ):
-            # add to tree
-            t_a.add_node(Node(t_new, q_new), n_close)  # add to tree
+            # Add node (sample) to tree A
+            t_a.add_node(Node(t_new, q_new), n_close)
 
+            # Check if opposite tree (B) has node near sample -> might connect?
             n_close_opposite = t_b.get_nearest_neighbor(Node(t_rnd, q_rnd), v_max)
-
             assert n_close_opposite is not None
 
-            # should we steer here, instead of attempting to connect?
+            # TODO: should we steer here, instead of attempting to connect? (Valentin)
+            # Try to link both trees in one shot might fail. Steer more conservative,
+            # but more likely to succeed in cluttered spaces.
 
+            # Check if edge connecting tree A (node q_new) and B (node n_close_opposite) is collision-free 
             if edge_collision_free_with_moving_obs(
                 env,
                 q_new,
@@ -1724,22 +1581,20 @@ def plan_in_time_space_bidirectional(
             ):
                 logger.info("found a path")
 
-                # extract path from first tree
+                # Extract path from first tree A (q_new -> parent -> ... -> root_of_t_a)
                 configurations = [q_new.state()]
                 times = [t_new]
 
                 p = n_close
-
                 while p.parent is not None:
-                    configurations.append(p.q.state())
+                    configurations.append(p.q.state()) 
                     times.append(p.t)
-
                     p = p.parent
 
-                configurations.append(p.q.state())
+                configurations.append(p.q.state()) # Add root
                 times.append(p.t)
 
-                # extract path from other tree
+                # Extract path from other tree B (n_close_opposite -> parent -> ... -> root_of_t_b)
                 other_configurations = []
                 other_times = []
 
@@ -1747,75 +1602,74 @@ def plan_in_time_space_bidirectional(
                 while p.parent is not None:
                     other_configurations.append(p.q.state())
                     other_times.append(p.t)
-
                     p = p.parent
 
                 other_configurations.append(p.q.state())
                 other_times.append(p.t)
 
-                # print("BBBBBBBB")
-                # print(times[-1])
-                # print(other_times[-1])
-                # print(times[0])
-                # print(other_times[0])
-
-                configurations = configurations[::-1]
+                configurations = configurations[::-1] # Reverse first path
                 times = times[::-1]
 
-                path = configurations + other_configurations
+                path = configurations + other_configurations # Concatenate
                 times = times + other_times
 
+                # If t_a was forward_tree -> path already correct order
+                # If t_a was backward_tree -> path backward -> needs flip
                 if t_a.reverse:
                     path = path[::-1]
                     times = times[::-1]
 
-                # insert endtimes of robots
+                # Handle multi-robot different end times
+                # Sampled times may not include these robot-specific end times -> add them!
                 if len(robots) > 1:
                     for r in robots:
                         t_robot_end = end_times[r]
                         if t_robot_end > t0:
-                            # find time interval that contains t_robot_end
+                            # Find time interval that contains t_robot_end
                             for k in range(len(times)-1):
                                 if times[k] <= t_robot_end <= times[k+1]:
-                                    # linearly interpolate
+                                    # Found interval, linearly interpolate
                                     alpha = (t_robot_end - times[k]) / (times[k+1] - times[k])
                                     interpolated_pose = path[k] + alpha * (path[k+1] - path[k])
                                     p = interpolated_pose
                                     break
-
+                            
+                            # Projects that interpolated configuration back onto robot's preplanned path
                             p_conf = q0.from_flat(p)
                             p_proj = project_sample_to_preplanned_path(t_robot_end, p_conf).state()
-
+                            # Append new point and time to the path
                             path.append(p_proj)
                             times.append(t_robot_end)
                     
-                    # sort according to increasing times
+                    # Sort according to increasing times
                     sorted_indices = np.argsort(times)
                     path = [path[i] for i in sorted_indices]
                     times = [times[i] for i in sorted_indices]
                     
-                # print(times)
-
-                # return path
+                # Return path
                 return TimedPath(time=times, path=path)
 
     print("Did not find a path in max_iters.")
     return None
 
-
 def shortcut_with_dynamic_obstacles(
-    env: BaseProblem, other_paths: MultiRobotPath, robots, path, max_iter=500
+    env: BaseProblem, 
+    other_paths: MultiRobotPath, 
+    robots, 
+    path, 
+    max_iter=500
 ):
+    """
+    
+    """
+    # Step 1: setup 
     logger.info("shortcutting")
-
-    # print((path.time[0]))
-    # print((path.time[-1]))
 
     ql = []
     offset = 0
     for r in env.robots:
         dim = env.robot_dims[r]
-        if r in robots:
+        if r in robots: # Only robots involved in this task
             ql.append(env.get_start_pos().state()[offset : offset + dim])
         offset += dim
 
@@ -1825,15 +1679,13 @@ def shortcut_with_dynamic_obstacles(
     def arr_to_config(q):
         return tmp_conf.from_flat(q)
 
-    # new_path = copy.copy(path)
-
+    # Step 2: discretization 
     discretized_path = []
     discretized_time = []
 
-    # discretize path
+    # Discretize path: add many intermediate points to TimedPath
     resolution = 0.1
-    for i in range(len(path.path) - 1):
-        # print('interpolating at index', i)
+    for i in range(len(path.path) - 1): # For each segment
         q0 = arr_to_config(path.path[i])
         q1 = arr_to_config(path.path[i + 1])
 
@@ -1841,85 +1693,75 @@ def shortcut_with_dynamic_obstacles(
         t1 = path.time[i + 1]
 
         dist = config_dist(q0, q1)
-        N = int(dist / resolution)
+        N = int(dist / resolution) # Number of intermediate points
         N = max(1, N)
 
-        # if len(robots) > 1:
-        #     print(t0)
-        #     coll_free = env.is_collision_free_np(path.path[i], None)
-        #     env.show()
-
-        for j in range(N):
+        for j in range(N): # Interpolate the N points (configurations and time)
             q = []
-            for k in range(q0.num_agents()):
-                qr = q0[k] + (q1[k] - q0[k]) / N * j
+            for k in range(q0.num_agents()): # For each robot (independently because #DOFs may vary)
+                qr = q0[k] + (q1[k] - q0[k]) / N * j # Linear interpolation
                 q.append(qr)
 
             discretized_path.append(np.concatenate(q))
 
-            t = t0 + (t1 - t0) * j / N
+            t = t0 + (t1 - t0) * j / N # Linear time interpolation
             discretized_time.append(t)
 
-            # if len(robots) > 1:
-            #     print(t)
-            #     coll_free = env.is_collision_free_np(np.concatenate(q), None)
-            #     env.show()
-
-    discretized_path.append(path.path[-1])
+    discretized_path.append(path.path[-1]) # Add final waypoint (loop only goes to N-1)
     discretized_time.append(path.time[-1])
 
+    # Discretized new path (dense)
     new_path = TimedPath(time=discretized_time, path=discretized_path)
 
-    num_indices = len(new_path.path)
-    end_times = other_paths.get_end_times(robots)
-
-    # print(end_times)
-
-    # start_time = time.time()
-
-    # if len(robots) > 1:
-    #     for k in range(len(new_path.path)):
-    #         print(new_path.time[k])
-    #         pose = env.start_pos.from_flat(new_path.path[k])
-    #         env.is_collision_free(pose, env.start_mode)
-    #         env.show(new_path.time[k] > 70)
-    #         time.sleep(0.01)
+    # Step 3: setup for shortcutting
+    num_indices = len(new_path.path) # Number of discretized waypoints
+    end_times = other_paths.get_end_times(robots) # When each robot finished prev task
 
     indices = {}
     offset = 0
+    # Build index map (which inidices in joint config belong to which robot?)
     for r in robots:
         dim = env.robot_dims[r]
-        indices[r] = np.arange(offset, offset + dim)
+        indices[r] = np.arange(offset, offset + dim) # 'rA': [0,1,2], 'rB': [3,4]
         offset += dim
 
+    # Step 4: MAIN SHORTCUTTING LOOP
     attempted_shortcuts = 0
     max_attempts = max_iter * 10
     for _ in range(max_attempts):
         if attempted_shortcuts > max_iter:
             break
-
-        i = np.random.randint(0, num_indices)
-        j = np.random.randint(0, num_indices)
+        
+        # 1: Shortcut selection (sample random 2 waypoints)
+        i = np.random.randint(0, num_indices) # Start of shortcut
+        j = np.random.randint(0, num_indices) # End of shortcut
 
         if i > j:
             i, j = j, i
 
         if abs(j - i) < 2:
-            continue
-
+            continue # Skip if i,j too close, e.g., when i=10 and j=11
+        
+        # Select random robot to optimize (one at a time)
         robot_idx_to_shortcut = np.random.randint(0, len(robots))
         robot_name_to_shortcut = robots[robot_idx_to_shortcut]
 
-        # if robot_name_to_shortcut == "a1":
-        #     continue
-
-        # we skip this attempt of shortcutting if the attempt tries to shortcut a path before the end time of this robot
+        # 2: End time (time at which robot finished prev task) filter
+        # Skip attempt of shortcutting if it tries shortcutting path before the end time of this robot
+        # Example:
+        #   rA end_time=5 rB end_time=10
+        #   t0=min(5,10)=5
+        #   Joint path waypoints: t=5   t=6  t=7   ...   t=10   ...   t=20
+        #                         rA moves freely        both moving
+        #                         rB holds position      toward task goal
+        #   Can't shortcut rB before t=10 because previous task not done!
         if (
             new_path.time[i] < end_times[robot_name_to_shortcut]
             or new_path.time[j] < end_times[robot_name_to_shortcut]
         ):
             continue
 
+        # 3: Cost improvement check
         assert new_path.time[i] < new_path.time[j]
 
         q0 = arr_to_config(new_path.path[i])
@@ -1928,47 +1770,47 @@ def shortcut_with_dynamic_obstacles(
         t0 = new_path.time[i]
         t1 = new_path.time[j]
 
-        # check if the shortcut improves cost
+        # Check if the shortcut improves cost
         if path_cost(
-            [q0.state(), q1.state()],
-            env.batch_config_cost,
-            agent_slices=q0._array_slice,
+            [q0.state(), q1.state()], env.batch_config_cost, agent_slices=q0._array_slice,
         ) >= path_cost(
             new_path.path[i:j], env.batch_config_cost, agent_slices=q0._array_slice
         ):
             continue
 
-        attempted_shortcuts += 1
+        attempted_shortcuts += 1 # Count it as valid attempt
 
+        # 4: Extract single-robot configurations (only check robot being shortcut)
         q0 = conf_type.from_list([q0[robot_idx_to_shortcut]])
         q1 = conf_type.from_list([q1[robot_idx_to_shortcut]])
 
-        # append paths that are not involved
-        if len(robots) > 1:
+        # 5: Multi-robot coordination setup
+        # Create paths for OTHER robots (not being shortcut)
+        if len(robots) > 1: # Multiple robots involved in this task
             tmp_other_paths = other_paths
 
-            # tmp_other_paths = copy.deepcopy(other_paths)
             tmp_paths = {}
             for r in robots:
-                if r != robot_name_to_shortcut:
-                    ind = indices[r]
-                    tmp_paths[r] = TimedPath(
+                if r != robot_name_to_shortcut: # OTHER robot
+                    ind = indices[r] # Indices of robot r in joint config
+                    tmp_paths[r] = TimedPath( # Extract robot r's configs from ALL waypoints
                         path=[pt[ind] * 1.0 for pt in new_path.path],
                         time=copy.deepcopy(discretized_time),
                     )
 
-            # uninvolved_indices = np.array([ind for r, ind in indices.items() if r != robots[robot_to_shortcut]]).flatten()
-            # tmp_path = TimedPath(time=discretized_time, path=[pt[uninvolved_indices] for pt in new_path.path])
+            # Add OTHER robot (its existing path from previous tasks)
+            # Temporary (just for collision check) 
             tmp_other_paths.add_path(
                 [r for r in robots if r != robot_name_to_shortcut],
                 Path(path=tmp_paths, task_index=3),
                 None,
                 is_escape_path=True,
             )
-        else:
-            tmp_other_paths = other_paths
+        else: # Single robot case 
+            tmp_other_paths = other_paths # No need to add anything
 
-        # this is wrong for partial shortcuts atm.
+        # TODO: what is meant by "partial shortcuts"?
+        # This is wrong for partial shortcuts atm. (Valentin)
         if edge_collision_free_with_moving_obs(
             env,
             q0,
@@ -1980,16 +1822,14 @@ def shortcut_with_dynamic_obstacles(
             end_times,
             resolution=env.collision_resolution,
         ):
-            # prev_time = new_path.time[i]
-            # prev_time_2 = new_path.time[j]
-            # print(robot_name_to_shortcut)
-            # print(prev_time, prev_time_2)
-            for k in range(j - i):
+            for k in range(j - i): # k = 0,1,2,.. (waypoints 1,2,3,..)
                 ql = []
                 for r_idx, r in enumerate(robots):
                     if r_idx == robot_idx_to_shortcut:
+                        # Interpolate shortcut (keep structure intact -> same #waypoints)
                         q = q0[0] + (q1[0] - q0[0]) / (j - i) * k
                     else:
+                        # Keep original path for OTHER
                         q = new_path.path[i + k][indices[r]] * 1.0
                     ql.append(q)
 
@@ -1998,47 +1838,11 @@ def shortcut_with_dynamic_obstacles(
                     new_path.time[j] - new_path.time[i]
                 )
 
-                # if len(robots) > 1:
-                #     res = env.is_collision_free_np(np.concatenate(ql), None)
-                #     if res == False:
-                #         print("AAAAAAAA")
-                #     env.show(True)
-                    # other_robots = [r for r in robots if r != robot_name_to_shortcut]
-                    # other_poses = tmp_other_paths.get_robot_poses_at_time(other_robots, new_path.time[i+k])
-
-                    # pose = []
-                    # for r in robots:
-                    #     if r == robot_name_to_shortcut:
-                    #         pose.append()
-
-
-            # assert new_path.time[i] == prev_time
-            # assert new_path.time[j] == prev_time_2
-
-            # if len(robots) > 1:
-            #     for k in range(len(new_path.path)):
-            #             env.show_config(env.start_pos.from_flat(new_path.path[k]), blocking=False)
-
+        # Cleanup: remove OTHER robot's path (was just temp added for collision check)
         if len(robots) > 1:
             tmp_other_paths.remove_final_escape_path(
                 [r for r in robots if r != robot_name_to_shortcut]
             )
-
-    # print(arr_to_config(path.path[0]).state())
-    # print(arr_to_config(new_path.path[0]).state())
-
-    # print(arr_to_config(path.path[-1]).state())
-    # print(arr_to_config(new_path.path[-1]).state())
-
-
-    # print((path.time[0]))
-    # print((new_path.time[0]))
-
-    # print((path.time[-1]))
-    # print((new_path.time[-1]))
-
-    # assert np.isclose(arr_to_config(path.path[0]).state(),arr_to_config(new_path.path[0]).state())[0]
-    # assert np.isclose(arr_to_config(path.path[-1]).state(),arr_to_config(new_path.path[-1]).state())[0]
 
     logger.info(
         f"original cost: {path_cost(path.path, env.batch_config_cost, agent_slices=q0._array_slice)}"
@@ -2051,7 +1855,6 @@ def shortcut_with_dynamic_obstacles(
     info = {}
 
     return new_path, info
-
 
 def plan_robots_in_dyn_env(
     config,
@@ -2066,7 +1869,11 @@ def plan_robots_in_dyn_env(
     t_lb=-1,
     use_bidirectional_planner=True,
 ) -> Tuple[Optional[Dict[str, TimedPath]], Optional[NDArray]]:
-    # plan
+    """
+    
+    """
+    # Call i) bidirectional or ii) unidirectional time-space RRT planner 
+    # Both consider the other_paths as dynamic obstacles
     if use_bidirectional_planner:
         path = plan_in_time_space_bidirectional(
             ptc, env, t0, other_paths, robots, end_times, goal, t_lb
@@ -2076,40 +1883,23 @@ def plan_robots_in_dyn_env(
             ptc, env, t0, other_paths, robots, end_times, goal, t_lb
         )
 
+    # Failure check (no collision free path, time budget, too many iterations)
     if path is None:
         print("Did not find a path in dyn env.")
         return None, None
 
-    # if len(robots) == 1:
-    #     plt.figure()
-
-    #     x = [pt[0] for pt in path.path]
-    #     y = [pt[1] for pt in path.path]
-
-    #     plt.plot(x, y)
-
-    # if len(robots) == 2:
-    #     plt.figure()
-
-    #     x = [pt[0] for pt in path.path]
-    #     y = [pt[1] for pt in path.path]
-
-    #     plt.plot(x, y)
-
-    #     x = [pt[3] for pt in path.path]
-    #     y = [pt[4] for pt in path.path]
-
-    #     plt.plot(x, y)
-
-    # postprocess
+    # Post-processing (shortcutting) of TimedPath (tuple with .time and .path)
     if config.shortcut_iters > 0:
         postprocessed_path, info = shortcut_with_dynamic_obstacles(
             env, other_paths, robots, path, max_iter=config.shortcut_iters
         )
         path = postprocessed_path
-        # display_multi_robot_path(env, path)
 
-    # take the separate paths apart
+    # Split joint path into per-robot paths
+    # Example: rA with DOF=3 and rB with DOF=2
+    # - Without split: 1 TimedPath with 5-nunmber configs
+    # - With split: 2 TimedPath with 3-number and 2-number configs
+    # Why? MultiRobotPath.add_path() expects per-robot paths.   
     separate_paths = {}
     offset = 0
     logger.debug(f"end times {end_times}")
@@ -2119,40 +1909,75 @@ def plan_robots_in_dyn_env(
         per_robot_times = []
         for i in range(len(path.path)):
             if path.time[i] >= end_times[r]:
-                per_robot_times.append(path.time[i])
-                c_n.append(path.path[i][offset : offset + dim])
+                per_robot_times.append(path.time[i]) # Extract per_robot time
+                c_n.append(path.path[i][offset : offset + dim]) # Extract per_robot path
 
-        offset += dim
+        offset += dim # Tracks where in joint config each robot's DOFs start
 
         separate_paths[r] = TimedPath(time=per_robot_times, path=c_n)
 
-    #     if len(robots) == 1:
-    #         x = [pt[0] for pt in separate_paths[r].path]
-    #         y = [pt[1] for pt in separate_paths[r].path]
+        # Example
+        # separate_paths = {
+        #   'rA': TimedPath(
+        #       time=[t1,t2,...],
+        #       path = [[x1,y1,theta1],[x2,y2,theta2],...]
+        #   ),
+        #   'rB': TimedPath(
+        #       time=[t1,t2,...],
+        #       path = [[x1,y1],[x2,y2],...]
+        #   ),
+        # } First 
 
-    #         plt.plot(x, y)
-    #         # plt.show()
+    # Return value 1: sepatarte_paths 
+    # - Used by main planner to build robot_path
+    # - Task planning: added with actual task_index and next_mode
+    # - Escape planning: added with task_index=-1 and next_mode=None
 
-    # if len(robots) == 2:
-    #     for r in robots:
-    #         x = [pt[0] for pt in separate_paths[r].path]
-    #         y = [pt[1] for pt in separate_paths[r].path]
-
-    #         plt.plot(x, y)
-    # plt.show()
-
+    # Return value 2: path.path[-1] 
+    # - Final joint config used by main planner to compute mode transition
+    
     return separate_paths, path.path[-1]
 
+"""
+NOTES (Liam):
+Remove later also comments within code, keep for now for better understanding 
+=============================================================================================
+PRIORITIZED PLANNER
 
+Plans for multiple robots by solving one task at a time in a fixed sequence order. 
+
+OUTER LOOP:
+- Iterates over task-sequence, keeping best solution found
+
+INNER LOOP:
+- Removes prior escape path from involved robots
+- Plans task path, treating all other committed robot paths as dynamic obstacles
+- Resolves the mode transition via a precomputed task_id_sequence
+- Commits the path, then plans escape paths to send involved robots to safe poses to not
+  block future tasks
+=============================================================================================
+"""
 @dataclass
 class PrioritizedPlannerConfig:
     # gamma: float = 0.7
     # distance_metric: str = "euclidean"
     use_bidirectional_planner: bool = True
-    shortcut_iters: int = 100
-    multirobot_shortcut_iters: int = 100
+    shortcut_iters: int = 100 # Per-task optimization (after planning each task)
+    multirobot_shortcut_iters: int = 100 # Global optimization (after completing all tasks)
 
 
+# TODO LIST
+# [x] PP: fallback when skill fails? -> return None, None 
+# [x] PP: skill terminal configuration should always enable the mode transition?
+# [x] PP: consistent bookkeeping of the per-robot skill ranges? -> with skill_time_ranges Dict
+# [o] SKill: instead check if current mode contains a skill for a given robot? to avoid having to mark it explicitly before
+# [x] Skill: move rollout functions to the skill classes
+# [o] Skill: add skill.duration attribute to deterministic timed skill class
+# [o] Skill: done() needs to be more generic terminal condition (check if current state is goal state), instead of convergence condition
+# [o] PP: What if we don't have another sequence when breaking to the outer loop?
+# [o] General: add types to arguments in all functions
+# [o] PP: multi-robot 
+# [o] Skill: consistent definitions 
 class PrioritizedPlanner(BasePlanner):
     def __init__(
         self,
@@ -2162,25 +1987,90 @@ class PrioritizedPlanner(BasePlanner):
         self.env = env
         self.config = config
 
+    # TODO (Liam)
+    def _execute_skill_task(
+            self, 
+            task, 
+            env, 
+            involved_robots, 
+            start_pose, 
+            t0,
+            prev_plans: MultiRobotPath,
+            end_times: Dict[str, float],
+    ):
+        """
+        Rollout skill
+        Return path in same format as plan_robots_in_dyn_env()
+        IMPORTANT: handles single robot skills for now!
+        """
+        skill = task.skill
+        robot = involved_robots[0] # Single robot assumed 
+        q_init = start_pose[0]
+
+        # Rollout the skill
+        result = skill.rollout(q_init, env, t0)
+        traj, times = result.trajectory, result.times
+        
+        # Goal check 
+        if not task.goal.satisfies_constraints(traj[-1], mode=None, tolerance=1e-3):
+            logger.warning("skill did not reach its goal")
+            return None, None
+
+        # Collision check
+        conf_type = type(env.get_start_pos())
+        for k in range(len(times)-1):
+            qs_conf = conf_type.from_list([traj[k]])
+            qe_conf = conf_type.from_list([traj[k+1]])            
+            if not edge_collision_free_with_moving_obs(
+                env,
+                qs_conf,            # Configuration object
+                qe_conf,            # Configuration object
+                times[k],
+                times[k+1],
+                prev_plans,
+                involved_robots,    
+                end_times,
+                env.collision_resolution,
+            ):
+                return None, None # Failure makes outer loop try different sequence
+
+        # Avoid shortcutting -> taken care of in robot_mode_shortcut()        
+
+        # Return 
+        timed_path = TimedPath(time=times.tolist(), path=list(traj)) # TODO (Liam) check if correct type!
+        path = {robot: timed_path}
+        final_pose = traj[-1]
+        return path, final_pose
+    
     def plan(
         self,
         ptc: PlannerTerminationCondition,
         optimize: bool = True,
     ) -> Tuple[List[State] | None, Dict[str, Any]]:
-        q0 = self.env.get_start_pos()
+        """
+        Try multiple task orderings / sequences (OUTTER LOOP), and for each one, plan tasks
+        one at a time, treating already-planned robots as dynamic obstacles (INNER LOOP).
+
+        Stop OUTTER LOOP if time runs out or sequences are exhausted.
+
+        Returns: 
+        - Best solution found (sequence of states with configs + modes)
+        - Info dictionary with costs, times, paths for each sequence tried?
+        """
+        # SETUP
+        q0 = self.env.get_start_pos() # Starting config and mode
         m0 = self.env.get_start_mode()
 
         conf_type = type(self.env.get_start_pos())
-
         robots = self.env.robots
 
-        # get a sequence for the tasks from the environment
-        # this is a constraint that this planner has, we need to have a sequence for planning
-
+        # Requirement check: prio-planner requires safe HOME poses for escape paths
         if self.env.spec.home_pose != SafePoseType.HAS_SAFE_HOME_POSE:
             raise ValueError("No safe home pose")
 
-        # life is easier if we just assume that we get a sequence-env for now
+        # Dependency check: assume that we get a sequence-env for now
+        # Option 1: Fixed sequence (returns [0,1,2,..])
+        # Option 2: Oracle provides sequence (tries to find best sequence)
         if (
             self.env.spec.dependency != DependencyType.FULLY_ORDERED
             and self.env.spec.dependency != DependencyType.UNORDERED
@@ -2189,69 +2079,77 @@ class PrioritizedPlanner(BasePlanner):
 
         sequence_cache = []
         skipped_sequences = 0
-
         info = {"costs": [], "times": [], "paths": []}
-
         best_path = None
-
         computation_start_time = time.time()
+
+        # OUTTER LOOP: Try different task sequences
         while True:
             if ptc.should_terminate(0, time.time() - computation_start_time):
-                break
-
-            robot_paths = MultiRobotPath(q0, m0, robots)
-
+                break # Time budget exhausted
+            
+            # Step 1: create a fresh plan 
+            robot_paths = MultiRobotPath(q0, m0, robots)  # Stores individual path segments for every robot
             seq_index = 0
-
             success = False
             env = copy.deepcopy(self.env)
-            env.C_cache = {}
+            env.C_cache = {} # Collision cache
 
-            # plan for a single sequence
-            sequence = env.get_sequence()
+            # Step 2: get task ordering from oracle/env (plan for single sequence)
+            sequence = env.get_sequence() 
             logger.info(f"Planning for sequence \n{sequence}")
 
-            # check if we planned for this sequence before
+            # Skip duplicates (if we planned for this sequence before)
             if sequence in sequence_cache:
                 logger.debug("Planned for this sequence before, skipping.")
                 skipped_sequences += 1
 
                 if skipped_sequences > 10:
-                    break
+                    break # Oracle is repeating
 
-                continue
+                continue # Try next sequence
 
             skipped_sequences = 0
             sequence_cache.append(sequence)
 
-            # construct task id sequence
+            # Step 3: construct task ID sequence 
             task_id_sequence = [env.start_mode.task_ids]
 
+            # Iterate through each task in the sequence
             for s in range(len(sequence) - 1):
-                ids = copy.deepcopy(task_id_sequence[-1])
+                ids = copy.deepcopy(task_id_sequence[-1]) # Copy last known task assignments
                 task_idx = sequence[s]
                 task = env.tasks[task_idx]
 
                 for robot_idx, r in enumerate(env.robots):
                     if r not in task.robots:
-                        continue
-
+                        continue # Skip robots not involved in this task
+                    
+                    # Find next task for this robot
                     for i in range(s + 1, len(sequence)):
                         next_task_idx = sequence[i]
                         next_task = env.tasks[next_task_idx]
 
+                        # Update their ID
                         if r in next_task.robots:
                             ids[robot_idx] = next_task_idx
-                            break
+                            break # Found next task for this robot
 
                 task_id_sequence.append(ids)
 
+                # Example:
+                # T0 for R0, T1 for R1, T2 for R0
+                # sequence = [0,1,2] 
+                # task_id_sequence = [(0,1), (2,1), (2,1)]
+                # Each entry is the next task each robot will execute 
+
+            # INNER LOOP: Plan tasks in sequence
             while True:
                 if ptc.should_terminate(0, time.time() - computation_start_time):
                     print("Terminating due to max time")
                     break
 
-                # get next active task
+                # Step 1: prepare current task to be planned from sequence
                 task_index = sequence[seq_index]
                 task = env.tasks[task_index]
                 involved_robots = task.robots
@@ -2261,148 +2159,127 @@ class PrioritizedPlanner(BasePlanner):
                 logger.info(f"robots: {involved_robots}")
                 logger.info(f"sequence index {seq_index}")
 
-                # figure out when this task can end at the earliest
+                # Figure out when this task can end at the earliest
                 earliest_end_time = robot_paths.get_final_non_escape_time()
                 logger.debug(f"earliest end time {earliest_end_time}")
 
-                # remove escape path from plan
-
+                # Discard previously assigned escape paths (will replan after task)
                 robot_paths.remove_final_escape_path(involved_robots)
 
                 end_times = robot_paths.get_end_times(involved_robots)
                 t0 = min([v for k, v in end_times.items()])
 
-                # get current robot position and last planned time
+                # Get exact time (t0) and config (start_pose) from which new plan for involved robots should start
                 logger.debug("Collecting start times")
                 start_time = min([t for _, t in end_times.items()])
                 start_pose = robot_paths.get_robot_poses_at_time(
                     involved_robots, start_time
                 )
 
-                # sample goal from the task
+                # Goal specification for this task
                 task_goal = task.goal
 
-                # plan actual task
-                current_time = time.time()
-                planning_ptc = RuntimeTerminationCondition(
-                    ptc.max_runtime_in_s - (current_time - computation_start_time)
-                )
+                # Step 2: motion planning or skill execution
+                # TODO (Liam) Rollout skill instead of planning
+                if hasattr(task, 'skill') and task.skill is not None:
+                    # Execute skill (instead of planning)
+                    path, final_pose = self._execute_skill_task(
+                        task, env, involved_robots, start_pose, t0,
+                        prev_plans=robot_paths,
+                        end_times=end_times,
+                    )
+                    if path is None:
+                        logger.warn('Skill trajectory in collision, trying next sequence.')
+                        break
+                        
+                else:
+                    # Planning with time aware RRT
+                    current_time = time.time()
+                    planning_ptc = RuntimeTerminationCondition( # Create time-limited termination condition
+                        ptc.max_runtime_in_s - (current_time - computation_start_time)
+                    )
 
-                # display_multi_robot_path(env, robot_paths)
-
-                path, final_pose = plan_robots_in_dyn_env(
-                    self.config,
-                    planning_ptc,
-                    env,
-                    t0,
-                    robot_paths,
-                    involved_robots,
-                    start_pose,
-                    end_times,
-                    task_goal,
-                    earliest_end_time,
-                    use_bidirectional_planner=self.config.use_bidirectional_planner,
-                )
+                    # PLAN THE TASK
+                    path, final_pose = plan_robots_in_dyn_env(
+                        self.config,
+                        planning_ptc,
+                        env,
+                        t0,
+                        robot_paths, # Previous paths (moving obstacles)
+                        involved_robots,
+                        start_pose,
+                        end_times,
+                        task_goal,
+                        earliest_end_time, # t_lb: lower bound for planning (task can't finish before)
+                        use_bidirectional_planner=self.config.use_bidirectional_planner,
+                    )
 
                 logger.debug(f"final_pose {final_pose}")
 
                 if path is None:
                     logger.warn("Failed planning.")
-                    break
+                    break # Sequence failed, try next sequence
 
-                # add plan to overall path
+                # Step 3: determine mode transition (next mode for task)
                 prev_mode = robot_paths.get_mode_at_time(
                     robot_paths.get_final_non_escape_time()
                 )
                 curr_mode = prev_mode
+
+                # Not reached terminal mode yet
                 if not env.is_terminal_mode(curr_mode):
                     final_time = path[involved_robots[0]].time[-1]
                     logger.debug(f"start_time {path[involved_robots[0]].time[0]}")
                     logger.debug(f"final_time: {final_time}")
+                    
                     q = []
                     for r in env.robots:
                         if r in involved_robots:
-                            q.append(path[r].path[-1])
+                            q.append(path[r].path[-1]) # Final pose of just-planned robot path
                         else:
-                            q.append(
+                            q.append( # Where the other robots are
                                 robot_paths.get_robot_poses_at_time([r], final_time)[0]
                             )
 
                     logger.debug(f"curr mode {curr_mode}")
                     logger.debug(f"final pose in path {final_pose}")
+
+                    # Get next modes that can be reached from current mode and config
                     next_modes = env.get_next_modes(
                         env.start_pos.from_list(q), curr_mode
                     )
+                    # If ambiguity -> use precomputed task_id_sequence
                     if len(next_modes) > 1:
                         for next_mode in next_modes:
                             if next_mode.task_ids == task_id_sequence[seq_index + 1]:
-                                break
-                            # for this_id, next_id in zip(curr_mode.task_ids, next_mode.task_ids):
-                            #     if this_id != next_id and next_id == sequence[seq_index+1]:
-                            #         break
+                                break # The mode matching the planned sequence
                     else:
                         assert len(next_modes) == 1
                         next_mode = next_modes[0]
                 else:
                     next_mode = curr_mode
 
+                # Commit path to robot_paths
                 robot_paths.add_path(
-                    involved_robots, Path(path=path, task_index=task_index), next_mode
+                    involved_robots,
+                    Path(path=path, task_index=task_index),
+                    next_mode,
                 )
 
-                # if len(involved_robots) > 1:
-                # display_multi_robot_path(env, robot_paths, blocking=False)
-
-
-                # for r in env.robots:
-                #     for p in robot_paths.paths[r]:
-                #         print(p.path)
-
-                # print("Drawing path")
-                # final_time = robot_paths.get_final_time()
-                # # plt.figure()
-                # for r in robots:
-                #     x = []
-                #     y = []
-                #     for t in np.arange(0, int(final_time), 0.1):
-                #         p = robot_paths.get_robot_poses_at_time([r], t)[0]
-                #         x.append(p[0])
-                #         y.append(p[1])
-
-                #     plt.plot(x, y)
-
-                # plt.show()
-
-                # for r in robots:
-                #     print("___________________")
-                #     print("R", r)
-                #     tprev = 0
-                #     for segment in robot_paths.paths[r]:
-                #         print("next segment")
-                #         print(segment.task_index)
-                #         for t in segment.path.time:
-                #             # if t < tprev:
-                #             print(t)
-
-                #             # tprev = t
-
-                # print("A")
-                # print(seq_index)
-                # print(env.sequence)
-                if seq_index + 1 >= len(sequence):
+                # Step 4: check if done
+                if seq_index + 1 >= len(sequence): # Check if this was the last task in the given sequence
                     success = True
                     logger.info("Found a solution.")
                     break
 
-                # plan escape path
-
+                # Step 5: plan escape path
                 logger.info("planning escape path")
-                escape_start_time = path[involved_robots[0]].time[-1]
+                escape_start_time = path[involved_robots[0]].time[-1] # End of main task path
                 end_times = robot_paths.get_end_times(involved_robots)
                 failed_escape_planning = False
 
                 for r in involved_robots:
-                    q_non_blocking = env.safe_pose[r]
+                    q_non_blocking = env.safe_pose[r] # Goal for escape path
                     escape_goal = SingleGoal(q_non_blocking)
 
                     escape_start_pose = robot_paths.get_robot_poses_at_time(
@@ -2416,13 +2293,14 @@ class PrioritizedPlanner(BasePlanner):
                         ptc.max_runtime_in_s - (current_time - computation_start_time)
                     )
 
+                    # Planning escape paths with time aware RRT
                     escape_path, _ = plan_robots_in_dyn_env(
                         self.config,
                         escape_planning_ptc,
                         env,
                         t0,
                         robot_paths,
-                        [r],
+                        [r], # Simpler single-robot planning problem
                         escape_start_pose,
                         end_times,
                         escape_goal,
@@ -2434,7 +2312,8 @@ class PrioritizedPlanner(BasePlanner):
                         failed_escape_planning = True
                         break
 
-                    # add plan to overall path
+                    # Commit escape path to involved robots
+                    # No mode change, just temporary repositioning move 
                     robot_paths.add_path(
                         [r],
                         Path(path=escape_path, task_index=-1),
@@ -2443,42 +2322,34 @@ class PrioritizedPlanner(BasePlanner):
                     )
 
                 if failed_escape_planning:
-                    break
-
-                # check if there is a task left to do
-                # print("escape path")
-                # display_multi_robot_path(env, robot_paths)
+                    break # Consider entire task sequence as failure
 
                 seq_index += 1
 
-            if False:
-                for k, v in robot_paths.paths.items():
-                    for p in v:
-                        print(p.path.time)
-                        print(p.path.path)
-
-                display_multi_robot_path(env, robot_paths)
-
-            # re-organize the data such that it is in the same format as before
+            # Step 6: finalizing a successful sequence
+            # Re-organize the data such that it is in the same format as before
+            # - Internal use: prioritized planner works with robot_paths (per-robot, separated paths) 
+            # - External use: unified path as list of State(config, mode) objects
             if success:
                 path = []
 
-                T = robot_paths.get_final_time()
-                N = 5 * int(np.ceil(T))
+                T = robot_paths.get_final_time() # Total makespan of the plan
+                N = 5 * int(np.ceil(T)) # Sample plan at N evenly spaced time points
+
+                # At each sample time, query configs and current modes from robot_paths
                 for i in range(N):
                     t = i * T / (N - 1)
-
                     q = robot_paths.get_robot_poses_at_time(env.robots, t)
                     config = conf_type.from_list(q)
                     mode = robot_paths.get_mode_at_time(t)
 
-                    state = State(config, mode)
-
+                    state = State(config, mode) # Pack information into State
                     path.append(state)
 
                 end_time = time.time()
                 cost = path_cost(path, env.batch_config_cost)
 
+                # Check and save if this path is the best
                 if len(info["costs"]) == 0 or info["costs"][-1] > cost:
                     info["costs"].append(cost)
                     info["times"].append(end_time - computation_start_time)
@@ -2488,11 +2359,19 @@ class PrioritizedPlanner(BasePlanner):
 
                     logger.info(f"Added path with cost {cost}.")
 
+                # Step 7: global shortcutting / post-optimization (on ENTIRE assembled path)
                 if self.config.multirobot_shortcut_iters > 0:
                     path_w_doubled_modes = []
+
+                    # Shortcutter needs to know exactly when a mode change happens
+                    # Example: State(qA,m1) -> State(qB,m2) ambiguous as we don't know where exactly change happens 
                     for i in range(len(path)):
                         path_w_doubled_modes.append(path[i])
 
+                        # At every mode transition, insert extra state State(old_config, new_mode)
+                        # That way, mode transition happens at exact same config 
+                        # TODO (Liam) but shouldn't this node already exist? (our sampled transition nodes...)
+                        # NO!! This we do in the composite PRM planner (when sampling transition nodes!)
                         if i + 1 < len(path) and path[i].mode != path[i + 1].mode:
                             path_w_doubled_modes.append(
                                 State(path[i].q, path[i + 1].mode)
@@ -2500,6 +2379,8 @@ class PrioritizedPlanner(BasePlanner):
 
                     path = path_w_doubled_modes
 
+                    # Iteratively tries shorcutting segments 
+                    # Respecting mode constraints (no mode shortcutting over modes, except for inactive (other) robots) 
                     shortcut_path, info_shortcut = robot_mode_shortcut(
                         env,
                         path,
