+++ test_notes.md
@@ -21,3 +21,13 @@ pytest tests/test_envs.py -v
 # VISUALIZE AVAILABLE ENVIRONMENTS
 python3 examples/show_problems.py --mode list_all
 python3 examples/show_problems.py rai.box_stacking --mode modes
+
+
+# VIDEO
+python3 examples/run_planner.py rai.single_agent_screw       --planner=prioritized --max_time=30 --optimize --save
+
+python3 examples/display_single_path.py ./out/20260220_224545_rai.single_agent_screw/prioritized/0/path_0.json rai.single_agent_screw --export
+
+cd z.vid
+
+ffmpeg -framerate 30 -i %04d.png -c:v libx264 -pix_fmt yuv420p out.mp4
